% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads-package.R
\docType{data}
\name{accepted}
\alias{accepted}
\alias{current}
\alias{infos}
\alias{previous}
\title{RnBeads option values and restrictions}
\format{\code{infos} is a \code{data.frame} containing information about all options in \pkg{RnBeads}. Row names in
        this table are the option names; the column names are \code{"Type"}, \code{"Named"}, \code{"Null"},
        \code{"Max"}, \code{"Min"}, \code{"MaxInclusive"} and \code{"MinInclusive"}.
        \code{accepted} is a \code{list} containing the sets of accepted values for some of the options.
        \code{current} is a \code{list} with current values for all options.
        \code{previous} is a \code{list} with previous values for the affected options; this list is only temporarily
        used while setting option values through \code{\link{rnb.options}} or \code{\link{rnb.xml2options}}.}
\description{
The values of options in RnBeads are stored in dedicated R objects accompanying the package. These objects are named
\code{infos}, \code{accepted}, \code{current} and \code{previous}. They should not be loaded or otherwise operated on
by users. Please refer to the documentation of \code{\link{rnb.options}} for accessing and modifying option values in
\pkg{RnBeads}.
}
\author{
Yassen Assenov
}
\keyword{datasets}

