% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{add.age.histogram}
\alias{add.age.histogram}
\title{add.age.histogram}
\usage{
\method{add}{age.histogram}(report, ages)
}
\arguments{
\item{report}{report object to be modified.}

\item{ages}{ages for which the histogram should be created.}
}
\value{
modified report object with the age histogram.
}
\description{
This function creates an age histogram for the predicted ages in the case
where there is no age annotation available for comparison with predicted ages.
}
\author{
Michael Scherer
}

