% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{add.agecomparison.plot}
\alias{add.agecomparison.plot}
\title{add.agecomparison.plot}
\usage{
\method{add}{agecomparison.plot}(report, object, actualAges, predictedAges)
}
\arguments{
\item{report}{Report in which the corresponding plots should be integrated.}

\item{object}{Methylation dataset as an object of type \code{\linkS4class{RnBeadSet}}.}

\item{actualAges}{Annotated ages as found in the sample annotation sheet}

\item{predictedAges}{Ages as predicted by the age prediction algorithm}
}
\value{
The modified report with the age plots integrated
}
\description{
Adds a plot that compares predicted ages by the age predictor with the annotated ages of the samples. One can compare
different traits with each other. Until now the following traits are (if annotated in the sample annotation sheet)
supported:	tissue, gender, hivstatus, sex, disease state, smoking status, ethnicity
}
\author{
Michael Scherer
}

