% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{add.combination.plot}
\alias{add.combination.plot}
\title{add.combination.plot}
\usage{
\method{add}{combination.plot}(report, object, actualAges, predictedAges)
}
\arguments{
\item{report}{Report in which the corresponding plot should be integrated.}

\item{object}{Methylation dataset as an object of type \code{\linkS4class{RnBeadSet}}.}

\item{actualAges}{Annotated ages as found in the sample annotation sheet}

\item{predictedAges}{Ages as predicted by the age prediction algorithm}
}
\value{
The modified report with the quantile plot integrated
}
\description{
This function adds a plot that is the combination of the upper two plots in one plot.The upper 
panel shows the distribution and the quantiles for the differences between predicted and annotated
ages and the lowe panel shows the corresponding errors
}
\author{
Michael Scherer
}

