% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\docType{methods}
\name{addDiffVarTable,RnBDiffMeth-method}
\alias{addDiffVarTable,RnBDiffMeth-method}
\alias{addDiffVarTable}
\title{addDiffVarTable-methods}
\usage{
\S4method{addDiffVarTable}{RnBDiffMeth}(object, var.table, comparison,
  grp.labs = c("group1", "group2"))
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}

\item{var.table}{Differential variability table to add}

\item{comparison}{character or index of the comparison of the table to retrieve}

\item{grp.labs}{character vector of length 2 specifying the names of the groups being compared}
}
\value{
the updated RnBDiffMeth object
}
\description{
Adds a differential variability table
}
\note{
Caveat: if disk dumping is enabled the resulting object tables will be stored in the initial location of the object.
}
\author{
Michael Scherer
}
