% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialVariability.R
\name{addReportPlot.diffVar.meth}
\alias{addReportPlot.diffVar.meth}
\title{addReportPlot.diffVar.meth}
\usage{
addReportPlot.diffVar.meth(report, var.table, meth.table, comparison.name,
  group.name1, group.name2)
}
\arguments{
\item{report}{Object of class \code{\linkS4class{Report}} to which the plot should be added.}

\item{var.table}{Table containing the differentially variable sites.}

\item{meth.table}{Table containing the differentially methylated sites.}

\item{comparison.name}{Name of the comparison to be conducted.}
}
\description{
This function compares differentially methylated sites (DMCs) with differentially variable sites (DMVs) in a scatterplot
and colors significantly differential sites with different colors.
}
