% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{age.anti.transformation}
\alias{age.anti.transformation}
\title{age.anti.transformation}
\usage{
age.anti.transformation(x, adult.age = 20)
}
\arguments{
\item{x}{input age about to be transformed back}

\item{adult.age}{threshold for deciding between 'fast' and 'slow' changes in DNA methylation}
}
\value{
age transformed back
}
\description{
This function the ages that were transformed by \code{age.transformation} back to actual ages.
}
\details{
The function is directly taken from Steve Horvath's prediction algorithm
DNA methylation age of human tissues and cell types, Steve Horvath, Genome Biology, 2013
}
\author{
Steve Horvath
}

