% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{agePredictor450}
\alias{agePredictor450}
\title{agePredictor450}
\usage{
agePredictor450(rnbSet, path)
}
\arguments{
\item{rnbSet}{An \code{RnBeadSet} object containing the relevant methylation infos.}

\item{path}{Path to a csv file in which a trained predictor should lay.
DEFAULT:    The predefined predictor is loaded.}
}
\value{
RnBeadSet with a new column in the phenotypic table in which the predicted ages
	are annotated
}
\description{
This function is the corresponding instance of the general age prediction algorithm for data
created by Illumina Infinium Bead Chips
}

