% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{agePredictorBiseq}
\alias{agePredictorBiseq}
\title{agePredictorBiseq}
\usage{
agePredictorBiseq(rnbSet, path)
}
\arguments{
\item{rnbSet}{a \code{RnBiseqSet} object containing the relevant methylation infos.}

\item{path}{path to a csv file in which a trained predictor should exist.}
}
\value{
modified \code{RnBiseqSet} with two new columns in the phenotypic table (predicted_ages
	for the predicted epigenetic ages and age_increase for the difference between predicted
	and annnotated ages)
}
\description{
This function is the corresponding instance of the general age prediction algorithm for data
created by a sequencing-based approach.
}
\author{
Michael Scherer
}

