% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotations.R
\name{append.cpg.stats}
\alias{append.cpg.stats}
\title{append.cpg.stats}
\usage{
append.cpg.stats(genome.data, regionlist)
}
\arguments{
\item{genome.data}{Genome of interest.}

\item{regionlist}{Genomic regions as a list of \code{GRanges} objects (or an obect of type \code{GRangesList}),
containing one set of regions per chromosome.}
}
\value{
The modified \code{regionlist}. Two columns are appedned to the metadata of each element in this list -
        \code{"CpG"} and \code{"GC"}. If the metadata already contains these columns, this function appends columns
        with similar names.
}
\description{
Appends additional metadata columns for CpG count and GC density to the specified regions.
}
\author{
Yassen Assenov
}

