% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeadRawSet-class.R
\name{as.RnBeadRawSet}
\alias{as.RnBeadRawSet}
\title{Conversion to/from RnBeadRawSet}
\description{
The \code{"as"} method can be used for the following conversions:
\itemize{
  \item{}{\code{MethyLumiSet} (in package \pkg{methylumi}) to \code{\linkS4class{RnBeadRawSet}}}
  \item{}{\code{RnBeadRawSet} to \code{MethyLumiSet}}
  \item{}{\code{RGChannelSet} (in package \pkg{minfi}) to \code{\linkS4class{RnBeadRawSet}}}
}
}

