% Generated by roxygen2 (4.0.2): do not edit by hand
\name{auto.select.rank.cut}
\alias{auto.select.rank.cut}
\title{auto.select.rank.cut}
\usage{
auto.select.rank.cut(p, r, alpha = 0.1)
}
\arguments{
\item{p}{vector of p-values}

\item{r}{vector of ranks}

\item{alpha}{the percentile to select the top and bottom part of the list}
}
\value{
the maximum rank fulfilling the criterion
}
\description{
automatically select a rank cutoff for given ranks and p-values
current implementation: sort the p-values according to rank. select as rank cutoff
the rank for which the worst (i.e. max) p-value in the top list is still smaller than
the best (i.e. min) p-value of the group of worst-ranking p-values of equal size as the top-list
}
\author{
Fabian Mueller
}

