% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterArchitectureLSF.R
\docType{class}
\name{ClusterArchitectureLSF-class}
\alias{ClusterArchitectureLSF-class}
\title{ClusterArchitectureLSF Class}
\description{
A child class of \code{\linkS4class{ClusterArchitecture}} implementing specifications of IBM LSF architectures.
}
\details{
Follow this template if you want to create your own ClusterArchitecture class.
}
\section{Slots}{

see \code{\linkS4class{ClusterArchitecture}}
}

\section{Methods}{

\describe{
  \item{\code{\link{getSubCmdTokens,ClusterArchitectureLSF-method}}}{Returns a vector of command line tokens corresponding to submitting
  a job with the given command to the cluster}
}
}

\author{
Michael Scherer
}
