% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ClusterArchitectureSGE-class}
\alias{ClusterArchitectureSGE-class}
\title{ClusterArchitectureSGE Class}
\description{
A child class of \code{\linkS4class{ClusterArchitecture}} implementing specifications of Sun Grid Engine (SGE) architectures.
}
\details{
Follow this template if you want to create your own ClusterArchitecture class.
}
\section{Slots}{

see \code{\linkS4class{ClusterArchitecture}}
}

\section{Methods}{

\describe{
  \item{\code{\link{getSubCmdTokens,ClusterArchitectureSGE-method}}}{Returns a vector of command line tokens corresponding to submitting
  a job with the given command to the cluster}
}
}
\author{
Fabian Mueller
}

