% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeadSet-class.R, R/RnBiseqSet-class.R
\name{coercion-methods}
\alias{coercion-methods}
\alias{coercion-methods}
\title{as("RnBeadSet", "MethyLumiSet")}
\description{
Convert a \code{\linkS4class{RnBeadSet}} object to \code{\linkS4class{MethyLumiSet}}

Convert a \code{\linkS4class{RnBeadSet}} object to a "mock" \code{\linkS4class{RnBiseqSet}} object
(used in the combine method)
}
