% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialVariability.R
\name{cols.to.rank.site}
\alias{cols.to.rank.site}
\alias{cols.to.rank.region}
\title{cols.to.rank.site}
\usage{
cols.to.rank.site(diff.var)

cols.to.rank.region(diff.var)
}
\arguments{
\item{diff.var}{A differential variability table.}
}
\value{
A matrix with the absolute values of the relevant columns
}
\description{
Return a matrix containing the negative absolute values of the information used to rank the sites. Those are currently:
the variance difference, the log ratio in variances and the p-value from the statistical test.
}
\author{
Michael Scherer
}
