% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{combine,RnBSet,RnBSet-method}
\alias{combine,RnBSet,RnBSet-method}
\alias{combine,RnBSet-method}
\title{combine-methods}
\usage{
\S4method{combine}{RnBSet,RnBSet}(x, y)
}
\arguments{
\item{x,y}{\code{\linkS4class{RnBeadSet}}, \code{\linkS4class{RnBeadRawSet}}
                    or \code{\linkS4class{RnBiseqSet}} object}
}
\value{
combined \code{\linkS4class{RnBeadSet}}, \code{\linkS4class{RnBeadRawSet}} or
\code{\linkS4class{RnBiseqSet}} object
}
\description{
Combine two objects inheriting from \code{\linkS4class{RnBSet}} class
}
\details{
The sample sets of \code{x} and \code{y} should be unique.
Sample annotation information is merged only for columns which have identical names in both objects.
CpG sites of the new object are a union of those present in both objects.
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
r1 <- rnb.set.example
r1 <- remove.samples(r1,samples(rnb.set.example)[1:5])
i <- which(r1@sites[,2] == 15 | r1@sites[,2] == 21)
sites.rem.r1 <- union(sample(1:nrow(meth(rnb.set.example)),500),i)
r1 <- remove.sites(r1,sites.rem.r1)
r2 <- rnb.set.example
r2 <- remove.samples(r2,samples(rnb.set.example)[6:12])
sites.rem.r2 <- sample(1:nrow(meth(rnb.set.example)),800)
r2 <- remove.sites(r2,sites.rem.r2)
rc <- combine(r1,r2)
#assertion: check the number of sites
sites.rem.c <- intersect(sites.rem.r1,sites.rem.r2)
(nrow(meth(rnb.set.example))-length(sites.rem.c)) == nrow(meth(rc))
}
}

