% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combine.diffMeth.objs}
\alias{combine.diffMeth.objs}
\title{combine.diffMeth.objs}
\usage{
combine.diffMeth.objs(obj.list)
}
\arguments{
\item{obj.list}{a list containing outputs from \code{rnb.run.differential}}
}
\description{
combine differential methylation objects (output from \code{rnb.run.differential}).
To be more precise, the \code{diffmeth} and \code{dm.enrich} are merged.
individual objects that are merged are assumed to belong to the same analysis
and vary only in their indexing of region types and comparisons
}
\author{
Fabian Mueller
}

