% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialVariability.R
\name{computeDiffVar.region}
\alias{computeDiffVar.region}
\title{computeDiffVar.region}
\usage{
computeDiffVar.region(var.table, regions2sites, includeCovg = FALSE)
}
\arguments{
\item{var.table}{A differential variability table on the site level}

\item{regions2sites}{mapping from those sites to the region of interest}

\item{includeCovg}{Flag indicating if data set contains coverage information.}
}
\value{
A differential variability table.
}
\description{
This function computes differnetially variable tables on the region level.
}
