% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialVariability.R
\name{computeDiffVar.default.site}
\alias{computeDiffVar.default.site}
\alias{computeDiffVar.site}
\alias{computeDiffVar.extended.site}
\title{computeDiffVar.site}
\usage{
computeDiffVar.default.site(meth.matrix, inds.g1, inds.g2,
  method = rnb.getOption("differential.variability.method"),
  adjustment.table = NULL, eps = 0.01)

computeDiffVar.extended.site(meth.matrix, inds.g1, inds.g2,
  method = rnb.getOption("differential.variability.method"),
  adjustment.table = NULL, covg = NULL,
  covg.thres = rnb.getOption("filtering.coverage.threshold"))
}
\arguments{
\item{meth.matrix}{Methylation matrix on which differential variability analysis should be conducted.}

\item{inds.g1}{Column indices of the first group.}

\item{inds.g2}{Column indices of the second group.}

\item{method}{Differrential analysis method to conduct. Possibilities are: "diffVar" and "iEVORA".}

\item{adjustment.table}{A \code{data.frame} containing variables to adjust for in the testing}
}
\value{
Differential Variability data frame for the analysis.
}
\description{
This function performs the differential variability analysis on the given methylatio matrix and the group assignment
specified by the column indices.
}
\author{
Michael Scherer
}
