% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{convert.string.ages}
\alias{convert.string.ages}
\title{convert.string.ages}
\usage{
convert.string.ages(age_information)
}
\arguments{
\item{age_information}{input ages}
}
\value{
the input ages in numeric format
}
\description{
This function converts different formats of age information into numeric format to
be used in the following functions.
}
\author{
Michael Scherer
}

