% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.hex.summary.plot}
\alias{create.hex.summary.plot}
\title{create.hex.summary.plot}
\usage{
create.hex.summary.plot(df2p, x = colnames(df2p)[1], y = colnames(df2p)[2],
  q = colnames(df2p)[3], bins = 128, fun = median, ...)
}
\arguments{
\item{df2p}{\code{data.frame} to be plotted.}

\item{x}{name of the variable in \code{df2p} considered as x-axis}

\item{y}{name of the variable in \code{df2p} considered as y-axis}

\item{q}{name of the variable in \code{df2p} considered as quantity to be summarized over bins}

\item{bins,fun,...}{arguments to be passed on to \code{stat_summary_hex}}
}
\value{
\code{ggplot} object
}
\description{
Creates a summary plot binning the data given by a certain quantity in heagonal bins
}
\author{
Fabian Mueller
}

