% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{cv.array}
\alias{cv.array}
\title{cv.array}
\usage{
cv.array(rnbSet)
}
\arguments{
\item{rnbSet}{\code{RnBeadSet} object on which the cross validation should be perfomed}
}
\value{
the result of the cross validation in a data.frame format
}
\description{
This function randomly changes the order of the samples in the data set and then
calls the \code{general.cv} to perform cross-validation for a RnBSet derived from
BeadChip arrays
}
\author{
Michael Scherer
}

