% Generated by roxygen2 (4.0.2): do not edit by hand
\name{densRanks}
\alias{densRanks}
\title{densRanks}
\usage{
densRanks(x, y = NULL, nbin = 128, bandwidth)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{nbin}{number of bins}

\item{bandwidth}{bandwidth}
}
\description{
Rank the points accordind to density of the region they fall in. Densities are computed
as Kernel Density estimates. The method and parameters are implemented in analogy to
\code{grDevices::densCols}
}
\author{
Fabian Mueller
}

