% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RnBSet-class.R, R/RnBeadRawSet-class.R, R/RnBeadSet-class.R
\docType{methods}
\name{destroy,RnBSet-method}
\alias{destroy}
\alias{destroy,RnBSet-method}
\alias{destroy,RnBeadRawSet-method}
\alias{destroy,RnBeadSet-method}
\title{destroy-methods}
\usage{
\S4method{destroy}{RnBSet}(object)

\S4method{destroy}{RnBeadRawSet}(object)

\S4method{destroy}{RnBeadSet}(object)
}
\arguments{
\item{object}{object inheriting from \code{\linkS4class{RnBSet}}}
}
\value{
Nothing of particular interest
}
\description{
Remove tables stored to disk from the file system. Useful for cleaning up disk dumped objects.
}

