% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{destroy,RnBDiffMeth-method}
\alias{destroy,RnBDiffMeth-method}
\title{destroy-methods}
\usage{
\S4method{destroy}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}
}
\value{
Nothing of particular interest
}
\description{
remove tables stored to disk from the file system. Useful for cleaning up disk dumped objects.
CAUTION: currently only works with reloaded objects
}
\author{
Fabian Mueller
}

