% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigFf.R
\docType{methods}
\name{dim,BigFfMat-method}
\alias{dim,BigFfMat-method}
\title{Dimensions of BigFfMat}
\usage{
\S4method{dim}{BigFfMat}(x)
}
\arguments{
\item{x}{BigFfMat object}
}
\description{
Dimensions of BigFfMat
}
