% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads-package.R
\docType{data}
\name{epic_v1_manifest}
\alias{epic_v1_manifest}
\title{XenoMix EPICv1 (hg38)}
\format{
\code{data.frame} provided by Illumina capturing information about the probe design.
}
\description{
This is the manifest file for EPICv1, which is required for detecting inter-species probes for estimating mouse contamination
with XenoMix
}
\author{
Nima Esmaeelpour
}
\keyword{datasets}
