% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{general.cv}
\alias{general.cv}
\title{general.cv}
\usage{
general.cv(fitFunction, ages, methData, k = 10)
}
\arguments{
\item{fitFunction}{a function that fits a predictor from training data to 
predict age from methylation data (e.g \code{simpeGlmnetEvaluate})}

\item{ages}{the ages to be trained on}

\item{methData}{input methylation matrix}

\item{k}{the fold parameter}
}
\value{
a matrix that contains the summarized quality measures for 
	the predictor which are:
		$cor[k+1]:	the mean correlation between predicted 
				ages and actual ages
		$cor[1:k]:	individual correlations between predicted 
				ages and actual ages for each fold
		$mean[k+1]:	the mean of the mean absolute deviation 
				between predicted ages and actual ages
		$mean[1:k]:	the individual mean absolute deviations for 
				each fold
		$median[k+1]:	the mean of the median absolute deviation
		$median[1:k]:	the individual median absolute devations 
				for each fold
}
\description{
This functions performs k-fold-cross-validation on the predictor with methylation 
data as input and ages to be trained on
}
\author{
Michael Scherer
}

