% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{geno,RnBSet-method}
\alias{geno,RnBSet-method}
\alias{geno}
\title{geno-methods}
\usage{
\S4method{geno}{RnBSet}(object, type = NULL)
}
\arguments{
\item{object}{dataset of interest.}

\item{type}{\code{character} singleton. Available options are \code{NULL} (all genotyping probes), \code{'nv'} for the
genotyping probes of the arrays, or \code{'rs'} for the SNP probes.}
}
\value{
\code{matrix} with genotyping information in the form of beta values. \code{0} stands for homozygous allele 1, \code{0.5}
     for heterozygous and \code{1} for homozygous allele 2.
}
\description{
Extracts genotyping information from either the SNP (rs) or mutation (nv) probes
}
