% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get.comparison.grouplabels,RnBDiffMeth-method}
\alias{get.comparison.grouplabels}
\alias{get.comparison.grouplabels,RnBDiffMeth-method}
\title{get.comparison.grouplabels-methods}
\usage{
\S4method{get.comparison.grouplabels}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}
}
\value{
character matrix containing comparison group names
}
\description{
Gets all comparison grouplabels represented in the object as character matrix of dimension n.comparisons x 2
where the columns specify group names 1 and 2 respectively
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
dm <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group","Treatment"))
get.comparison.grouplabels(dm)
}
}
\author{
Fabian Mueller
}

