% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\docType{methods}
\name{get.comparison.groupsizes,RnBDiffMeth-method}
\alias{get.comparison.groupsizes}
\alias{get.comparison.groupsizes,RnBDiffMeth-method}
\title{get.comparison.groupsizes-methods}
\usage{
\S4method{get.comparison.groupsizes}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}
}
\value{
character matrix containing comparison group sizes
}
\description{
Gets all comparison group sizes represented in the object as character matrix of dimension n.comparisons x 2
where the columns specify sizes of groups 1 and 2 respectively
}
\examples{
\donttest{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
dm <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group","Treatment"))
get.comparison.groupsizes(dm)
}
}
\author{
Fabian Mueller
}

