% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get.comparisons,RnBDiffMeth-method}
\alias{get.comparisons}
\alias{get.comparisons,RnBDiffMeth-method}
\title{get.comparisons-methods}
\usage{
\S4method{get.comparisons}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}
}
\value{
character vector containing comparisons
}
\description{
Gets all comparisons represented in the object as character vector
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
dm <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group","Treatment"))
get.comparisons(dm)
}
}
\author{
Fabian Mueller
}

