% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.covariates.ct}
\alias{get.covariates.ct}
\title{get.covariates.ct}
\usage{
get.covariates.ct(rnb.set)
}
\arguments{
\item{rnb.set}{\code{RnBSet} object}
}
\value{
an NxK matrix of K cell types contributions for N samples of the \code{rnb.set}. \code{NULL}
		   if the components have not been computed or added to \code{rnb.set}.
}
\description{
Retrieves an NxK matrix of cell type contributions stored in an RnBSet for a given target variable
}

