% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get.covg.thres,RnBDiffMeth-method}
\alias{get.covg.thres}
\alias{get.covg.thres,RnBDiffMeth-method}
\title{get.covg.thres-methods}
\usage{
\S4method{get.covg.thres}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{RnBDiffMeth object}
}
\value{
integer coverage threshold
}
\description{
Gets the coverage threshold employed for obtaining statistics in the differential methylation tables
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
dm <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group","Treatment"))
get.covg.thres(dm)
}
}
\author{
Fabian Mueller
}

