% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.cpg.stats}
\alias{get.cpg.stats}
\title{get.cpg.stats}
\usage{
get.cpg.stats(chrom.sequence, starts, ends)
}
\arguments{
\item{chrom.sequence}{Chromosome sequence, usually obtained from the assembly's genome definition. This must be an
object of type \code{MaskedDNAString}.}

\item{starts}{\code{integer} vector of start positions for the regions of interest.}

\item{ends}{\code{integer} vector of end positions for the regions of interest.}
}
\value{
Table of statistics for the regions in the form of a \code{matrix} with the following columns:
        \code{"CpG"} and \code{"GC"}. The columns contain the number of CpG dinucleoties and the number of C and G
        bases in each region.
}
\description{
Computes CpG-related statistics for the specified regions.
}
\author{
Yassen Assenov
}

