% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get.region.types,RnBDiffMeth-method}
\alias{get.region.types}
\alias{get.region.types,RnBDiffMeth-method}
\title{get.region.types-methods}
\usage{
\S4method{get.region.types}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}
}
\value{
character vector containing region types
}
\description{
Gets all region types represented in the object as character vector
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
dm <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group","Treatment"))
get.region.types(dm)
}
}
\author{
Fabian Mueller
}

