% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialVariability.R
\name{get.table.ids}
\alias{get.table.ids}
\title{Returns the colum names of the differential variability table.}
\usage{
get.table.ids(includeCovg = FALSE)
}
\arguments{
\item{includeCovg}{Flag indicating if dataset contains coverage information}
}
\value{
Column names of the differential variability table
}
\description{
Returns the colum names of the differential variability table.
}
