% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\docType{methods}
\name{get.variability.method,RnBDiffMeth-method}
\alias{get.variability.method,RnBDiffMeth-method}
\alias{get.variability.method}
\title{get.variability.method-methods}
\usage{
\S4method{get.variability.method}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{RnBDiffMeth object}
}
\value{
character describing the variability method
}
\description{
Gets the variability testing method used to obtain the p-values in the differential varibiality tables
}
\author{
Michael Scherer
}
