% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\docType{methods}
\name{get.variability.table,RnBDiffMeth-method}
\alias{get.variability.table,RnBDiffMeth-method}
\alias{get.variability.table}
\title{get.varibility.table-methods}
\usage{
\S4method{get.variability.table}{RnBDiffMeth}(object, comparison,
  region.type = "sites", undump = TRUE, return.data.frame = FALSE)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}

\item{comparison}{character or index of the comparison of the table to retrieve}

\item{region.type}{region type to be returned, defaults to 'sites'}

\item{undump}{Flag indicating whether to convert the table into a matrix instead of using the file descriptor.
Only meaningful if the if the objects's \code{disk.dump} slot is true.}

\item{return.data.frame}{should a data.frame be returned instead of a matrix?}
}
\value{
differential variability table. See \code{computeDiffVar.bin} for details.
}
\description{
Gets a differential variability table
}
\author{
Michael Scherer
}
