% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getModuleNumCores,RnBClusterRun-method}
\alias{getModuleNumCores}
\alias{getModuleNumCores,RnBClusterRun-method}
\title{getModuleNumCores-methods}
\usage{
\S4method{getModuleNumCores}{RnBClusterRun}(object)
}
\arguments{
\item{object}{\code{\linkS4class{RnBClusterRun}} object}
}
\value{
A named vector containing the number of cores for each module
}
\description{
Retrieves the number of cores used by each module
}
\author{
Fabian Mueller
}

