% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{getNumNaMeth,RnBSet-method}
\alias{getNumNaMeth,RnBSet-method}
\alias{getNumNaMeth}
\title{getNumNaMeth-methods}
\usage{
\S4method{getNumNaMeth}{RnBSet}(object, type = "sites", chunkSize = 1e+05, mask = NULL)
}
\arguments{
\item{object}{object inheriting from \code{\linkS4class{RnBSet}}}

\item{type}{"sites" or region type}

\item{chunkSize}{size of each submatrix (performance tuning parameter)}

\item{mask}{logical matrix. its entries will also be considered NAs in counting}
}
\value{
vector containing the number of NAs per site/region
}
\description{
for each site/region, the getNumNaMeth retrieves the number of NA values accross all samples.
Does this efficiently by breaking down the methylation matrix into submatrices
}
