% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getSubCmdStr,ClusterArchitecture-method}
\alias{getSubCmdStr}
\alias{getSubCmdStr,ClusterArchitecture-method}
\title{getSubCmdStr-methods}
\usage{
\S4method{getSubCmdStr}{ClusterArchitecture}(object, ...)
}
\arguments{
\item{object}{\code{\linkS4class{ClusterArchitecture}} object}

\item{...}{arguments passed on to \code{\link{getSubCmdTokens,ClusterArchitecture-method}}}
}
\value{
A string containing the submission command
}
\description{
Returns a string for the of command line corresponding to submitting
a job with the given command to the cluster.
}
\author{
Fabian Mueller
}

