% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getSubCmdTokens,ClusterArchitecture-method}
\alias{getSubCmdTokens}
\alias{getSubCmdTokens,ClusterArchitecture-method}
\title{getSubCmdTokens-methods}
\usage{
\S4method{getSubCmdTokens}{ClusterArchitecture}(object, cmd.tokens, log,
  job.name = "", res.req = character(0), depend.jobs = character(0))
}
\arguments{
\item{object}{\code{\linkS4class{ClusterArchitecture}} object}

\item{cmd.tokens}{a character vector specifying the executable command that should be wrapped in the cluster submission command}

\item{log}{file name and path of the log file that the submitted job writes to}

\item{job.name}{name of the submitted job}

\item{res.req}{character vector specifying required resources. The resource requirements should be the values of the vector,
the names should specify the resource name}

\item{depend.jobs}{character vector containg names or ids of jobs the submitted job will depend on.}
}
\value{
A character vector containing the submission command tokens
}
\description{
Returns a string for the of command line corresponding to submitting
a job with the given command to the cluster.
}
\details{
For a concrete child class implementation for a sun grid architecture specification see \code{\link{getSubCmdTokens,ClusterArchitectureSGE-method}}
}
\author{
Fabian Mueller
}

