% Generated by roxygen2 (4.0.2): do not edit by hand
\name{has.covariates.ct}
\alias{has.covariates.ct}
\title{has.covariates.ct}
\usage{
has.covariates.ct(rnb.set)
}
\arguments{
\item{rnb.set}{\code{RnBSet} object}
}
\value{
\code{TRUE} if the supplied object contains the cell type covariates information and \code{FALSE} otherwise
}
\description{
Checks whether the given \code{RnBSet} object contains cell type contribution estimates
}

