% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{hasCovg,RnBSet-method}
\alias{hasCovg}
\alias{hasCovg,RnBSet-method}
\title{hasCovg-methods}
\usage{
\S4method{hasCovg}{RnBSet}(object, type = "sites")
}
\arguments{
\item{object}{\code{RnBSet} of interest.}

\item{type}{\code{character} singleton. If \code{sites} or a region type summarized in the object}
}
\value{
\code{TRUE} if the \code{RnBSet} object contains coverage information for sites or the specified region type. \code{FALSE} otherwise
}
\description{
Returns \code{TRUE} if the \code{RnBSet} object contains coverage information for sites or the specified region type.
}
\examples{
\donttest{
library(RnBeads.hg19)
data(small.example.object)
## per-site beta-value matrix
hasCovg(rnb.set.example)
}
}

