% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\docType{methods}
\name{hasVariability,RnBDiffMeth-method}
\alias{hasVariability,RnBDiffMeth-method}
\alias{hasVariability}
\title{hasVariability-methods}
\usage{
\S4method{hasVariability}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{RnBDiffMeth object}
}
\value{
\code{TRUE} if the differential methylation object contains differential variability information. \code{FALSE} otherwise
}
\description{
Returns \code{TRUE} if the differential methylation object contains differential variability information
}
\author{
Michael Scherer
}
