% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{imputeBiseq}
\alias{imputeBiseq}
\title{imputeBiseq}
\usage{
imputeBiseq(methData)
}
\arguments{
\item{methData}{methylation data as a data frame or matix to be imputed}
}
\value{
the changed methylation data frame
}
\description{
Since the k-nearest-neighboring approach does not converge for all CpGs present in
sequencing data, a simple mean imputation is performed for this kind of data.
}
\author{
Michael Scherer
}

