% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{imputeRRBS}
\alias{imputeRRBS}
\title{imputeRRBS}
\usage{
imputeRRBS(methData)
}
\arguments{
\item{methData}{Methylation data as a data frame or matix to be imputed}
}
\value{
the changed methyaltion data frame
}
\description{
Since the k-nearest-neighbroing approach does not converge for all CpGs present in
sequencing data, a simple mean imputation is performed for this kind of data
}
\author{
Michael Scherer
}

