% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\docType{methods}
\name{includes.sites,RnBDiffMeth-method}
\alias{includes.sites}
\alias{includes.sites,RnBDiffMeth-method}
\title{includes.sites-methods}
\usage{
\S4method{includes.sites}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{RnBDiffMeth object}
}
\value{
\code{TRUE} if the differential methylation object contains site-level information. \code{FALSE} otherwise
}
\description{
Returns \code{TRUE} if the differential methylation object contains site-level information
}
\examples{
\donttest{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
dm <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group","Treatment"))
includes.sites(dm)
}
}
\author{
Fabian Mueller
}

