% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterArchitecture.R
\docType{methods}
\name{initialize,ClusterArchitecture-method}
\alias{initialize,ClusterArchitecture-method}
\title{initialize.ClusterArchitecture}
\usage{
\S4method{initialize}{ClusterArchitecture}(.Object,
  name = "ClusterArchitecture")
}
\arguments{
\item{.Object}{New instance of \code{ClusterArchitecture}.}

\item{name}{A name or identifier}
}
\description{
Initialize an ClusterArchitecture object
}
\author{
Fabian Mueller
}

