% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterArchitectureLSF.R
\docType{methods}
\name{initialize,ClusterArchitectureLSF-method}
\alias{initialize,ClusterArchitectureLSF-method}
\title{initialize.ClusterArchitectureLSF}
\usage{
\S4method{initialize}{ClusterArchitectureLSF}(.Object, name = "ClusterArchitectureLSF", ...)
}
\arguments{
\item{.Object}{New instance of \code{ClusterArchitectureLSF}.}

\item{name}{A name or identifier}

\item{...}{arguments passed on to the constructor of \code{\linkS4class{ClusterArchitecture}} (the parent class)}
}
\description{
Initialize an ClusterArchitecture object for a LSF
}
\author{
Michael Scherer
}
