% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterArchitectureSLURM.R
\docType{methods}
\name{initialize,ClusterArchitectureSLURM-method}
\alias{initialize,ClusterArchitectureSLURM-method}
\title{initialize.ClusterArchitectureSLURM}
\usage{
\S4method{initialize}{ClusterArchitectureSLURM}(.Object,
  name = "ClusterArchitectureSLURM", ...)
}
\arguments{
\item{.Object}{New instance of \code{ClusterArchitectureSLURM}.}

\item{name}{A name or identifier}

\item{...}{arguments passed on to the constructor of \code{\linkS4class{ClusterArchitecture}} (the parent class)}
}
\description{
Initialize an ClusterArchitecture object for a SLURM
}
\author{
Michael Scherer
}
