% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCluster.R
\docType{methods}
\name{initialize,RnBClusterRun-method}
\alias{initialize,RnBClusterRun-method}
\title{initialize.RnBClusterRun}
\usage{
\S4method{initialize}{RnBClusterRun}(.Object, architecture)
}
\arguments{
\item{.Object}{New instance of \code{RnBClusterRun}.}

\item{architecture}{A \code{\linkS4class{ClusterArchitecture}} object managing the settings for a scientific compute cluster.}
}
\description{
Initialize an RnBClusterRun object
}
\author{
Fabian Mueller
}

