% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\docType{methods}
\name{initialize,RnBDiffMeth-method}
\alias{initialize,RnBDiffMeth-method}
\title{initialize.RnBDiffMeth}
\usage{
\S4method{initialize}{RnBDiffMeth}(.Object,
  site.test.method = rnb.getOption("differential.site.test.method"),
  variability.method = rnb.getOption("differential.variability.method"),
  covg.thres = rnb.getOption("filtering.coverage.threshold"),
  disk.dump = FALSE, disk.path = NULL)
}
\arguments{
\item{.Object}{New instance of \code{RnBDiffMeth}.}

\item{site.test.method}{method which was applied to obtain the site-level p-values.}

\item{variability.method}{method to be used to calculate differentially variable sites. Has to be one of: `diffVar' or 
`iEVORA'.}

\item{covg.thres}{coverage threshold. Important for certain columns of the differential methylation tables.
See \code{computeDiffMeth.bin.site} and \code{computeDiffMeth.bin.region} for details.}

\item{disk.dump}{Flag indicating whether the tables should be stored on disk rather than in the main memory}

\item{disk.path}{Path on the disk for DMTs.Only meaningful if \code{disk.dump} is \code{TRUE}}
}
\description{
Initialize an RnBDiffMeth object
}
\author{
Fabian Mueller
}
