% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads-package.R
\docType{data}
\name{interspecies_probes_v1}
\alias{interspecies_probes_v1}
\title{XenoMix interspeciesprobes (v1)}
\format{
\code{list} of CGIs for the inter-species probes.
}
\description{
These are the inter-species probes for EPICv1
}
\author{
Nima Esmaeelpour
}
\keyword{datasets}
