% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{isImputed,RnBSet-method}
\alias{isImputed,RnBSet-method}
\alias{isImputed}
\title{isImputed}
\usage{
\S4method{isImputed}{RnBSet}(object)
}
\arguments{
\item{object}{Object for which the information should be returned}
}
\value{
TRUE, if the object has been imputed and FALSE otherwise.
}
\description{
Getter for the imputation field. Return TRUE, if the object has been imputed and FALSE otherwise.
}
\author{
Michael Scherer
}
