% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.region.subsegment.annotation}
\alias{load.region.subsegment.annotation}
\title{load.region.subsegment.annotation}
\usage{
load.region.subsegment.annotation(rnb.set, annotation.dir)
}
\arguments{
\item{rnb.set}{The \code{RnBSet object} with subsegments specified in the regions}

\item{annotation.dir}{a directory to load the annotation from. (binary \code{RData} format.)}
}
\value{
invisible \code{TRUE}
}
\description{
For the region annotation of a given \code{RnBSet} object. Subdivide each region into subsegments by
hierarchical clustering on the site distances in a particular region and then splitting the region into subregions consisting
of these site clusters. The number of clusters is determined in such way that the mean number of sites per cluster
is given by the \code{ns} parameter.
}
\author{
Fabian Mueller
}

