% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.rnb.diffmeth}
\alias{load.rnb.diffmeth}
\title{load.rnb.diffmeth}
\usage{
load.rnb.diffmeth(path)
}
\arguments{
\item{path}{path of the saved object (a directory containing a corresponding \code{rnbDiffMeth.RData} file and possibly \code{rnbDiffMeth_tables} files)}
}
\value{
the loaded \code{\linkS4class{RnBDiffMeth}} object
}
\description{
load a saved \code{\linkS4class{RnBDiffMeth}} object from disk
}
\author{
Fabian Mueller
}

