% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.rnb.set}
\alias{load.rnb.set}
\title{load.rnb.set}
\usage{
load.rnb.set(path, temp.dir = tempdir())
}
\arguments{
\item{path}{full path of the file or directory. If \code{archive} is \code{FALSE})
without an extension.}

\item{temp.dir}{\code{character} singleton which specifies temporary directory, used while loading}
}
\value{
Loaded object
}
\description{
Loading of the \code{RnBSet} objects with large matrices of type \pkg{ff}.
}

