% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logger.isinitialized}
\alias{logger.isinitialized}
\title{logger.isinitialized}
\usage{
logger.isinitialized()
}
\value{
\code{TRUE} if the logger was initialized and is in use; \code{FALSE} otherwise.
}
\description{
Checks if the logger is initialized.
}
\examples{
\dontrun{
if (!logger.isinitialized())
  logger.start(fname = NA)
}
}
\author{
Yassen Assenov
}
\seealso{
\code{\link{logger.start}} for initializing a logger or starting a section
}

