% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeads-package.R
\docType{data}
\name{lump.hg38}
\alias{lump.hg38}
\title{LUMP Support (hg38)}
\format{\code{lump.*} is a \code{list} of non-empty \code{integer} matrices, one per supported platform. Here, only `CpG` is 
         available for BS datasets.}
\description{
Those are the same sites as reported in \code{lump.hg19}, but lifted to `hg38` with UCSC's liftOver functionality. This only
applies for the CpG-wise sites; i.e. those used for sequencing data sets, since `hg38` is not supported for array-based data
sets.
}
\author{
Michael Scherer
}
\keyword{datasets}
