% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{mask.sites.meth,RnBSet-method}
\alias{mask.sites.meth,RnBSet-method}
\alias{mask.sites.meth}
\title{mask.sites.meth-methods}
\usage{
\S4method{mask.sites.meth}{RnBSet}(object, mask, verbose = FALSE)
}
\arguments{
\item{object}{Dataset of interest.}

\item{mask}{logical matrix indicating which sites should be masked}

\item{verbose}{if \code{TRUE} additional diagnostic output is generated}
}
\value{
The modified dataset.
}
\description{
Given a logical matrix, sets corresponding entries in the methylation table to NA (masking).
Low memory footprint
}
