% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{nsites,RnBSet-method}
\alias{nsites}
\alias{nsites,RnBSet-method}
\title{nsites-methods}
\usage{
\S4method{nsites}{RnBSet}(object, type = "sites")
}
\arguments{
\item{object}{\code{RnBSet} of interest.}

\item{type}{\code{character} singleton. If this is set to \code{"sites"} (default), the number of sites is returned.
Otherwise, this should be one of region types for for which the number of regions is returned.}
}
\value{
\code{integer} stating the number of sites/regions. \code{NA} if the regions have not been summarized yet.
}
\description{
Returns the number of sites/regions for a given \code{RnBSet} object
}
\examples{
\donttest{
library(RnBeads.hg19)
data(small.example.object)
nsites(rnb.set.example)
}
}
\seealso{
\code{\link[=meth,RnBSet-method]{meth}} Retrieving the matrix of methylation values
}

