% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{off,Report-method}
\alias{off}
\alias{off,Report-method}
\alias{off,ReportGgPlot-method}
\alias{off,ReportPlot-method}
\title{off-methods}
\usage{
\S4method{off}{Report}(.Object)

\S4method{off}{ReportPlot}(.Object)

\S4method{off}{ReportGgPlot}(.Object, handle.errors = FALSE)
}
\arguments{
\item{.Object}{Object to be closed.}

\item{handle.errors}{Flag indicating if the method should attempt to catch and process errors (e.g. I/O errors)
                     internally. Setting this to \code{TRUE} does not guarantee that the method never stops with an
                     error.}
}
\value{
The closed object, invisibly.
}
\description{
Performs cleanup and/or other other finishing activities and closes the specified device, connection, or document.
}

