% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report-methods.R, R/ReportPlot-methods.R
\docType{methods}
\name{off,Report-method}
\alias{off,Report-method}
\alias{off}
\alias{off,ReportPlot-method}
\alias{off,ReportGgPlot-method}
\title{off-methods}
\usage{
\S4method{off}{Report}(object, ...)

\S4method{off}{ReportPlot}(object)

\S4method{off}{ReportGgPlot}(object, handle.errors = FALSE)
}
\arguments{
\item{object}{Object to be closed.}

\item{...}{additional arguments}

\item{handle.errors}{Flag indicating if the method should attempt to catch and process errors (e.g. I/O errors)
internally. Setting this to \code{TRUE} does not guarantee that the method never stops with an
error.}
}
\value{
The closed object, invisibly.
}
\description{
Performs cleanup and/or other finishing activities and closes the specified device, connection, or document.
}
