% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parallel.disable}
\alias{parallel.disable}
\title{parallel.disable}
\usage{
parallel.disable()
}
\description{
Disables parallel processing.
}
\examples{
\dontrun{
parallel.getNumWorkers()
parallel.setup(2)
parallel.getNumWorkers()
parallel.disable()
parallel.getNumWorkers()
}
}
\author{
Fabian Mueller
}

