% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelProcessing.R
\name{parallel.isEnabled}
\alias{parallel.isEnabled}
\title{parallel.isEnabled}
\usage{
\method{parallel}{isEnabled}()
}
\value{
\code{TRUE} if multicore processing is enabled, \code{FALSE} otherwise.
}
\description{
Checks if whether parallel processing is enabled.
}
\examples{
\donttest{
parallel.isEnabled()
parallel.setup(2)
parallel.isEnabled()
parallel.teardown()
parallel.isEnabled()
}
}
\author{
Fabian Mueller
}

