% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelProcessing.R
\name{parallel.setup}
\alias{parallel.setup}
\title{parallel.setup}
\usage{
\method{parallel}{setup}(...)
}
\arguments{
\item{...}{Parameters for \code{registerDoParallel} from the \pkg{doParallel} package.
This allows, for instance, for specificying the number of workers.}
}
\value{
\code{TRUE} (invisible) to indicate that parallelization is set up.
}
\description{
Sets up parallel processing. Requires the \pkg{foreach} and \pkg{doParallel} packages
}
\note{
Requires the packages \pkg{foreach} and \pkg{doParallel}.
}
\examples{
\donttest{
parallel.setup(2)
parallel.teardown()
}
}
\author{
Fabian Mueller
}

