% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\docType{methods}
\name{prepare.diff.var,RnBDiffMeth-method}
\alias{prepare.diff.var,RnBDiffMeth-method}
\alias{prepare.diff.var}
\title{prepare.diff.var-methods}
\usage{
\S4method{prepare.diff.var}{RnBDiffMeth}(object,
  variability.method = rnb.getOption("differential.variabilit.method"),
  disk.dump = FALSE, disk.path.DMVs = NULL)
}
\arguments{
\item{object}{RnBDiffMeth object}

\item{variability.method}{Character describing the variability method to be used, Should be one of 'diffVar' or 'iEVORA'.}

\item{disk.dump}{Flag indicating if tables should be stored in disk.}

\item{disk.path.DMVs}{Path to store the differential varibility tables on disk.}
}
\value{
The modified RnBDiffMeth object
}
\description{
Prepares the RnBDiffMeth object to now also contain variability information.
}
\author{
Michael Scherer
}
