% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{qc,RnBeadSet-method}
\alias{qc}
\alias{qc,RnBeadSet-method}
\title{qc-methods}
\usage{
\S4method{qc}{RnBeadSet}(object)
}
\arguments{
\item{object}{Dataset of interest.}
}
\value{
Quality control information available for the dataset in the form of a \code{list} with two elements:
\code{Cy3} and \code{Cy5}.
}
\description{
Extracts HumanMethylation quality control information
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
qcinf<-dpval(rnb.set.example, row.names=TRUE)
head(qcinf$Cy3)
head(qcinf$Cy5)
}
}

