% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.geo.parse.characteristics_ch1}
\alias{read.geo.parse.characteristics_ch1}
\title{read.geo.parse.characteristics_ch1}
\usage{
read.geo.parse.characteristics_ch1(phenoData)
}
\arguments{
\item{phenoData}{Parsed phenotypic data frame object as output by \code{getGEO}.}
}
\value{
Phenotypic data frame with parsed sample information instead of \code{characteristics_ch1}.
}
\description{
Parses the sample information in all of the \code{characteristics_ch1} columns from a phenoData data frame as
obtained from \code{getGEO}.
}
\author{
Fabian Mueller
}
\seealso{
\code{\link[GEOquery]{getGEO}} in package \pkg{GEOquery}
}

