% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.geo}
\alias{read.geo}
\title{read.geo}
\usage{
read.geo(accession = NULL, filename = NULL,
  verbose = logger.isinitialized(), destdir = tempdir(),
  parse.characteristics_ch1 = TRUE)
}
\arguments{
\item{accession}{Character string representing the GEO series for download and parsing. It must start
with \code{"GSE"}.}

\item{filename}{File name of a previously downloaded GEO series matrix file or its gzipped
representation (in which case the filename must end in \code{".gz"}). Other file
formats, such as SOFT files, are not supported. Exactly one of \code{accession} or
\code{filename} must be specified.}

\item{verbose}{Flag indicating if messages should be created informing about the progress. If the
logger is initialized prior to calling this function, the informative messages are
sent to the logger. Warnings and errors are not affected by this parameters, the
function always outputs them.}

\item{destdir}{The destination directory for any downloads. Defaults to the
(architecture-dependent) temporary directory. Keep in mind that GEO series can be
demanding in terms of storage space.}

\item{parse.characteristics_ch1}{Flag indicating if additional sample characteristics (if such exist) are to be
parsed from the downloaded series matrix file(s).}
}
\value{
\code{\linkS4class{RnBeadSet}} object with phenotypic and beta value information; \code{NULL} if the given
        series contain no Infinium450K samples.
}
\description{
Imports Infinium 450K data series from the Gene Expression Omnibus.
}
\author{
Yassen Assenov
}
\seealso{
\code{getGEO} in package \code{GEOquery}
}

