% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.sample.annotation}
\alias{read.sample.annotation}
\title{read.sample.annotation}
\usage{
read.sample.annotation(fname, sep = rnb.getOption("import.table.separator"))
}
\arguments{
\item{fname}{Name of text file that contains a sample annotation table with a header. This method handles a variety
of file formats, including comma-separated values file exported from Genome Studio.}

\item{sep}{One-element \code{character} used as field separator in the tables file.}
}
\value{
Sample annotation table in the form of a \code{data.frame}, in which every row corresponds to a sample, and
        every column - to a trait.
}
\description{
Reads Illumina Infinium sample annotation.
}
\examples{
\dontrun{
  TODO add system annotation example
  annotation.file<-system.file("")
  sa<-read.sample.annotation(annotation.file)
  sa
}
}
\author{
Pavlo Lutsik
}

