% Generated by roxygen2 (4.0.2): do not edit by hand
\name{refFreeEWASP}
\alias{refFreeEWASP}
\title{refFreeEWASP}
\usage{
refFreeEWASP(X, inds.g1, inds.g2 = -inds.g1, adjustment.table = NULL,
  paired = FALSE, nboot = 100, ignore.na = TRUE,
  rescale.residual = TRUE)
}
\arguments{
\item{X}{Matrix on which the test is performed for every row}

\item{inds.g1}{column indices of group 1 members}

\item{inds.g2}{column indices of group 2 members}

\item{adjustment.table}{a \code{data.frame} containing variables to adjust for in the testing}

\item{paired}{should a paired analysis model be used. If so, the first index in \code{inds.g1} must correspond to the first
index in  \code{inds.g2} and so on.}

\item{nboot}{The number of bootstrapping resamples}

\item{ignore.na}{in this case all \code{NA} containing rows are removed}

\item{rescale.residual}{rescale the residual matrix as z-scores}
}
\value{
vector of p-values for the "adjusted" regression coefficients from the Reference-free EWAS model
}
\description{
Applies the reference-free cell-type heterogeneity adjustment model from [1] and
returns corrected p-values
}
\note{
Requires the package \pkg{RefFreeEWAS}.
}
\author{
Pavlo Lutsik
}
\references{
1. Houseman, E. Andres, John Molitor, and Carmen J. Marsit. "Reference-Free Cell Mixture Adjustments in Analysis of DNA Methylation Data."
Bioinformatics (2014): btu029.
}

