% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{regionMapping,RnBSet-method}
\alias{regionMapping}
\alias{regionMapping,RnBSet-method}
\title{regionMapping-methods}
\usage{
\S4method{regionMapping}{RnBSet}(object, region.type)
}
\arguments{
\item{object}{Dataset as an object of type inheriting \code{\linkS4class{RnBSet}}.}

\item{region.type}{region type. see \code{\link{rnb.region.types}} for possible values}
}
\value{
A list containing for each region the indices (as integers) of sites that belong to that region
}
\description{
get the mapping of regions in the RnBSet object to methylation site indices in the RnBSet object
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
promoter.probe.list <- regionMapping(rnb.set.example,"promoters")
#get the number of CpGs per promoter in the dataset:
sapply(promoter.probe.list,length)
}
}
\author{
Fabian Mueller
}

