% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{remove.regions,RnBSet-method}
\alias{remove.regions}
\alias{remove.regions,RnBSet-method}
\title{remove.regions-methods}
\usage{
\S4method{remove.regions}{RnBSet}(object, region.type)
}
\arguments{
\item{object}{Dataset of interest.}

\item{region.type}{Type of the region annotation for which the summarization should be removed}
}
\value{
object of the same class as the supplied one without the summarized methylation information for the specified region type
}
\description{
Remove the summarized methylation information for a given region type from an \code{RnBSet} object.
}
\examples{
\donttest{
library(RnBeads.hg19)
data(small.example.object)
summarized.regions(rnb.set.example)
rnb.set.reduced<-remove.regions(rnb.set.example, "genes")
summarized.regions(rnb.set.reduced)
}
}

