% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ReportGgPlot-class}
\alias{ReportGgPlot-class}
\alias{initialize,ReportGgPlot-method}
\title{ReportGgPlot Class}
\description{
Information about the files created to store one generated plot in a report. Report plots are initialized using the
function \code{\link{createReportGgPlot}}. It inherits from the \code{\linkS4class{ReportPlot}} class and handling is
analogous, except that it contains an additional slot to store a \code{ggplot} object.
}
\section{Slots}{

\describe{
  \item{\code{ggp}}{\code{ggplot} object to be printed}
}
}

\section{Notes}{

No device is being opened until \code{off(reportGgPlot)} is called.
}
\author{
Fabian Mueller
}

