% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.add.tables}
\alias{rnb.add.tables}
\title{rnb.add.tables}
\usage{
rnb.add.tables(report, tables, setting.names, selected.table = 1L,
  indent = 2L, ...)
}
\arguments{
\item{report}{Report to write the text to.}

\item{tables}{Non-empty \code{list} of tables, each one represented by a \code{\link{data.frame}} or
\code{\link{matrix}}. The names of this list are used as table identifiers; each one consists
of elements separated by underscore character (\code{_}).}

\item{setting.names}{List of table name element descriptors. Every variable elements in the table names must be
included in this list.}

\item{selected.table}{Index of the table to be initially selected in this listing.}

\item{indent}{Default indentation, in number of tabulation characters, to apply to every table.}

\item{...}{Other parameters passed to \code{\link{rnb.add.table}}.}
}
\value{
The modified report.
}
\description{
Generates HTML code for a listing of tables (of which only one is visible at any moment) in the specified report.
}
\author{
Yassen Assenov
}
\seealso{
\code{\link{rnb.add.table}} for adding a single table to a report; \code{\linkS4class{Report}} for other
  functions adding contents to an HTML report
}

