% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{rnb.bed.from.segmentation}
\alias{rnb.bed.from.segmentation}
\title{rnb.bed.from.segmentation}
\usage{
rnb.bed.from.segmentation(rnb.set, sample.name, type = "final",
  store.path = getwd())
}
\arguments{
\item{rnb.set}{An \code{\link{RnBSet-class}} object obtained by executing \code{rnb.execute.segmentation}.}

\item{sample.name}{The sample name for which segmentation was computed.}

\item{type}{The type of segmentation (\code{PMDs}, \code{UMRs}, \code{LMRs}, \code{HMDs} or \code{final}).}

\item{store.path}{Path to which the BED file is to be stored.}
}
\description{
This function creates a BED file from the segmentation result of \code{rnb.execute.segmentation} and stores it on disk.
}
\author{
Michael Scherer
}
