% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{rnb.boxplot.from.segmentation}
\alias{rnb.boxplot.from.segmentation}
\title{rnb.boxplot.from.segmentation}
\usage{
rnb.boxplot.from.segmentation(rnb.set, sample.name, type = "PMDs",
  store.path = getwd())
}
\arguments{
\item{rnb.set}{An \code{\link{RnBSet-class}} object obtained by executing \code{rnb.execute.segmentation}.}

\item{sample.name}{The sample name for which segmentation was computed.}

\item{type}{The type of segmentation (\code{PMDs}, \code{UMRsLMRs}, \code{HMRs} or \code{final}).}
}
\value{
An object of type \code{ggplot} visualizing the methylation values in the segments.
}
\description{
This function creates a boxplot from the segmentation result of \code{rnb.execute.segmentation}.
}
\author{
Michael Scherer
}
