% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.call.destructor}
\alias{rnb.call.destructor}
\title{rnb.call.destructor}
\usage{
rnb.call.destructor(object, ...)
}
\arguments{
\item{object}{object to be destroyed}

\item{...}{further arguments to the method \code{\link[=destroy,RnBSet-method]{destroy}}}
}
\value{
invisible \code{TRUE}
}
\description{
calls the destructor of an RnBSet, RnBeadSet or RnBeadRawSet object
conditionally on whether the \code{enforce.destroy.disk.dumps} option is enabled.
}
\author{
Fabian Mueller
}

