% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining.R
\name{rnb.combine.arrays}
\alias{rnb.combine.arrays}
\title{Combine array-based datasets}
\usage{
rnb.combine.arrays(dataset1, dataset2)
}
\arguments{
\item{dataset1}{First input dataset as an object of type inheriting \code{\linkS4class{RnBeadSet}}.}

\item{dataset2}{Second input dataset as an object of type inheriting \code{\linkS4class{RnBeadSet}}.}
}
\value{
Combined dataset as an object of type inheriting \code{\linkS4class{RnBeadSet}}.
}
\description{
Concatenates two array-based datasets focusing on the common probes.
}
\author{
Yassen Assenov
}
