% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining.R
\name{rnb.combine.seq}
\alias{rnb.combine.seq}
\title{rnb.combine.seq}
\usage{
rnb.combine.seq(x, y, type = "common")
}
\arguments{
\item{x}{An object of type \code{\link{RnBiseqSet-class}} used for concatenation}

\item{y}{Another object of type \code{\link{RnBiseqSet-class}} used for concatenation}

\item{type}{A character representing the type of combination. Needs to be one of \code{"common"},
\code{all.x}, \code{all.y} or \code{all}.}
}
\value{
An \code{\link{RnBiseqSet-class}} object with combined information
}
\description{
Initial implementation of the combine method for sequencing datasets.
}
\details{
The type parameters determines the mode of combination:
          \itemize{
            \item{\code{"common"}}{ The intersection between the sites present in the two 
                            datasets is used for the new dataset.}
            \item{\code{"all.x"}}{ All sites present in x are used.}
            \item{\code{"all.y"}}{ All sites present in y are used.}
            \item{\code{"all"}}{ The union between the sites of both datasets is used.}
          }
}
