% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{rnb.execute.age.prediction}
\alias{rnb.execute.age.prediction}
\title{rnb.execute.age.prediction}
\usage{
rnb.execute.age.prediction(object)
}
\arguments{
\item{object}{a \code{\linkS4class{RnBeadSet}} object for which age prediction should
be performed}
}
\value{
the modified  \code{\linkS4class{RnBeadSet}} object
}
\description{
Adds a column called 'predicted_ages' to the phenotypic table, in which the predicted ages of
the corresponding sample reside
}
\author{
Michael Scherer
}

