% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.execute.clustering.all}
\alias{rnb.execute.clustering.all}
\title{rnb.execute.clustering.all}
\usage{
rnb.execute.clustering.all(rnb.set)
}
\arguments{
\item{rnb.set}{Methylation dataset as an object of type inheriting \code{\linkS4class{RnBSet}}.}
}
\value{
List of list of clustering results; each element corresponds to one region type and is a list of objects
        of type \code{\linkS4class{RnBeadClustering}}.
}
\description{
Performs hierarchical clustering on the samples of the given dataset using multiple distance metrics and
agglomeration methods for all suggested site and region types.
}
\author{
Fabian Mueller
}
\seealso{
\code{\link{rnb.execute.clustering}} for performing clustering using a single site or region type.
}

