% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.execute.filter.summary}
\alias{rnb.execute.filter.summary}
\title{rnb.execute.filter.summary}
\usage{
rnb.execute.filter.summary(old.set, new.set)
}
\arguments{
\item{old.set}{Methylation dataset before filtering as an object of type inheriting \code{\linkS4class{RnBSet}}.}

\item{new.set}{Methylation dataset after filtering as an object of type inheriting \code{\linkS4class{RnBSet}}.}
}
\value{
\code{matrix} summarizing the number of removed and retained sites, samples, and (optionally) reliable and
        unreliable measurements.
}
\description{
Calculates a table summarizing the effect of the applied filtering procedures.
}
\details{
This function expects that the sites and samples in \code{new.set} are subsets of the sites and samples in
\code{old.set}, respectively. If this is not the case, it exists with an error.
}
\author{
Yassen Assenov
}
\seealso{
\code{\link{rnb.run.preprocessing}} for running the whole preprocessing module
}

