% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomewide.R
\name{rnb.execute.genomewide}
\alias{rnb.execute.genomewide}
\title{Genome-wide methylation level}
\usage{
rnb.execute.genomewide(dataset)
}
\arguments{
\item{dataset}{Methylation dataset to study, provided as an object of type inheriting \code{RnBSet}.}
}
\value{
\code{vector} of values in the range \code{[0, 1]}, storing the average beta values per sample.
}
\description{
Computes genome-wide methylation levels per sample.
}
\author{
Yassen Assenov
}
