% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.execute.high.coverage.removal}
\alias{rnb.execute.high.coverage.removal}
\title{rnb.execute.high.coverage.removal}
\usage{
rnb.execute.high.coverage.removal(rnb.set)
}
\arguments{
\item{rnb.set}{Methylation dataset as an object of type inheriting \code{\linkS4class{RnBiseqSet}}.}
}
\value{
\code{list} of two elements:
        \describe{
          \item{\code{"dataset"}}{The (possibly) modified dataset after retaining sites on autosomes only.}
          \item{\code{"filtered"}}{\code{integer} vector storing the indices of all removed sites.}
        }
}
\description{
Removes methylation sites with a coverage larger than 100 times the 95-percentile of coverage in each sample.
}
\author{
Fabian Mueller
}

