% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{rnb.execute.imputation}
\alias{rnb.execute.imputation}
\title{rnb.execute.imptutation}
\usage{
rnb.execute.imputation(rnb.set, method = rnb.getOption("imputation.method"),
  ...)
}
\arguments{
\item{rnb.set}{Dataset object inheriting from \code{\linkS4class{RnBSet}}.}

\item{method}{Imputation method to be used, must be one of \code{"mean.cpgs"}, \code{"mean.samples"},
\code{"random"}, \code{"knn"} or \code{"none"}.}

\item{...}{Optional arguments passed to knn.imputation}
}
\value{
The modified rnb.set object without missing methylation values.
}
\description{
Removes missing methylation values in the methylation matrix of the given object
}
\author{
Michael Scherer
}
