% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{rnb.final.segmentation}
\alias{rnb.final.segmentation}
\title{rnb.final.segmentation}
\usage{
rnb.final.segmentation(rnb.set, sample.name)
}
\arguments{
\item{rnb.set}{The \code{\link{RnBSet-class}}-object, for which segmentation was computed using \code{\link{rnb.execute.segmentation}}}

\item{sample.name}{The sample name in \code{rnb.set} for which segmentation was conducted.}
}
\description{
This function creates a single segmentation, assigning each region to PMD, HMR, UMR or LMR.
}
\author{
Michael Scherer
}
