% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.find.relative.site.coord}
\alias{rnb.find.relative.site.coord}
\title{rnb.find.relative.site.coord}
\usage{
rnb.find.relative.site.coord(rnb.set, region.type, extend.by = 0.33)
}
\arguments{
\item{rnb.set}{RnBSet object}

\item{region.type}{Region type for which the coordinates are computed}

\item{extend.by}{A number between 0 and 1 specifying the percentage by which a region is
extended in order to capture methylation information before region start and after region end}
}
\value{
a data frame containing the site index, the assigned region index and the relative coordinate
        The relative coordinate is 0 if the site's coordinate is identical to the region start coordinate
        and 1 if identical to the regions end coordinate and scaled inbetween. Coordinates can be
        less than 0 or larger than 1 if a site is in the upstream or downstream flanking region respectively
}
\description{
given a region types, assigns sites to regions and determines relative positions
of sites in the assigned region
}
\author{
Fabian Mueller
}

