% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{rnb.get.XY.shifts}
\alias{rnb.get.XY.shifts}
\title{rnb.get.XY.shifts}
\usage{
rnb.get.XY.shifts(rnb.set, signal.type = "raw")
}
\arguments{
\item{rnb.set}{Dataset of interest. Currently only \code{\linkS4class{RnBeadRawSet}} is supported.}

\item{signal.type}{Matrix or matrices to use in calculating the shifts. Currently ignored.}
}
\value{
Calculated shifts in the form of a matrix in which every row corresponds to a sample in \code{rnb.set} and
        the columns denote the shifts in the X and Y chromosomes.
}
\description{
Calculates the increase of signal on the sex chromosomes in the given dataset.
}
\author{
Yassen Assenov
}
