% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rnb.is.option}
\alias{rnb.is.option}
\title{rnb.is.option}
\usage{
rnb.is.option(txt)
}
\arguments{
\item{txt}{Potential option name. This should be a one-element \code{character} vector.}
}
\value{
\code{TRUE} if the specified parameter is a valid analysis option name; \code{FALSE} otherwise.
}
\description{
Checks if the specified text is an option name.
}
\examples{
\donttest{
rnb.is.option("logging") # TRUE
rnb.is.option("Logging") # FALSE
}
}
\author{
Yassen Assenov
}
\seealso{
\code{\link{rnb.options}} for getting and setting option values
}

