% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.load.sitelist}
\alias{rnb.load.sitelist}
\title{rnb.load.sitelist}
\usage{
rnb.load.sitelist(fname, verbose = FALSE)
}
\arguments{
\item{fname}{File listing the identifiers, one per line.}

\item{verbose}{Flag indicating if messages are to be printed. If the values is \code{TRUE} and a logger is
initialized, this function adds a message to the log.}
}
\value{
The loaded list of identifiers, or \code{NULL} if \code{fname} could not be open.
}
\description{
Loads a list of probe or site identifiers. This function is used in the preprocessing module for loading a whitelist
and/or a blacklist of identifiers.
}
\author{
Yassen Assenov
}
\seealso{
\code{\link{logger.start}} for initializing a logger
}

