% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{rnb.message.plot}
\alias{rnb.message.plot}
\title{rnb.message.plot}
\usage{
rnb.message.plot(txt)
}
\arguments{
\item{txt}{Text to be plotted.}
}
\value{
The newly initialized \code{ggplot} instance.
}
\description{
Creates a plot, using \pkg{ggplot2}, with a single text message.
}
\examples{
\donttest{
x11(width = 5, height = 5)
rnb.message.plot("Missing data")
}
}
\author{
Yassen Assenov
}

