% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.mval2beta}
\alias{rnb.mval2beta}
\title{rnb.mval2beta}
\usage{
rnb.mval2beta(mvals)
}
\arguments{
\item{mvals}{\code{numeric} vector or matrix of M values to be transformed.}
}
\value{
The calculated beta values.
}
\description{
Transforms M values to beta values.
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
mvals <- rnb.beta2mval(meth(rnb.set.example))
bvals <- rnb.mval2beta(mvals)
all((bvals-meth(rnb.set.example))<1e-10)
}
}
\author{
Pavlo Lutsik
}

