% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.options2xml}
\alias{rnb.options2xml}
\title{rnb.options2xml}
\usage{
rnb.options2xml(pretty = TRUE)
}
\arguments{
\item{pretty}{Flag indicating if the document should be formatted to be easily readable. For example, if this is set
to \code{TRUE} (default), every element is located on separate line. Formatting does not affect the
validity of the generated XML tree.}
}
\value{
XML document in the form of a \code{character} that encodes all options and their current values.
}
\description{
Exports all option values to an XML document.
}
\examples{
\dontrun{
cat(rnb.options2xml(), file = "rnbeads_options.xml")
}
}
\author{
Yassen Assenov
}

