% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rnb.performance.profile}
\alias{rnb.performance.profile}
\title{rnb.performance.profile}
\usage{
rnb.performance.profile(data.type = "450k", profile)
}
\arguments{
\item{data.type}{Type of dataset targeted; this must be one of \code{"450k"} (default) or \code{"bs"}.}

\item{profile}{Option profile; this must be one of \code{"minimal"}, \code{"moderate"} or \code{"full"}.}
}
\value{
Invisibly, a \code{list} containing the previous values of all modified options.
}
\description{
Enables one of the pre-installed anlaysis option profiles.
}
\author{
Pavlo Lutsik
}

