% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.plot.biseq.coverage.violin}
\alias{rnb.plot.biseq.coverage.violin}
\title{rnb.plot.biseq.coverage.violin}
\usage{
rnb.plot.biseq.coverage.violin(rnbs.set, samples, fname = NULL,
  type = "sites", covg.range = NULL, ...)
}
\arguments{
\item{rnbs.set}{RnBiseqSet object}

\item{samples}{unique sample identifiers. In case \code{rnb.getOption("identifiers.column")}
is not \code{NULL}, \code{samples} should attain values from the corresponding column,
or \code{colnames(meth(rnb.set))} otherwise}

\item{fname}{base filename for the files to be plotted. If NULL, the plot will not be written to file}

\item{type}{\code{character} singleton. If \code{site} the coverage
information is plotted for each methylation site. Otherwise
should be one of the regions returned by \code{rnb.region.types}}

\item{covg.range}{Vector of length 2 specifying the range of coverage to be plotted. if \code{NULL} (default) the entire range will be plotted}

\item{...}{other arguments to \code{\link{createReportPlot}}}
}
\value{
plot as an object of type \code{\linkS4class{ReportPlot}} if \code{writeToFile} is \code{TRUE} and of class
			\code{\link{ggplot}} otherwise.
}
\description{
Plots the violin plots of the coverage distribution
}
\author{
Fabian Mueller
TODO add examples
}

