% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.plot.ct.heatmap}
\alias{rnb.plot.ct.heatmap}
\title{rnb.plot.ct.heatmap}
\usage{
rnb.plot.ct.heatmap(ct.obj, type = "nonnegative", writeToFile = FALSE, ...)
}
\arguments{
\item{ct.obj}{Object of class \code{CellTypeInferenceResult} as returned by \link{rnb.execute.ct.estimation}.}

\item{type}{Type of cell type contributions to plot.}

\item{writeToFile}{If \code{TRUE}, the plot will be written to a file.}

\item{...}{Other arguments passed to \code{\link{createReportPlot}}.}
}
\value{
if \code{writeToFile=TRUE} an object of class \code{\linkS4class{ReportPlot}},
						or the protted matrix otherwise
}
\description{
Plot contributions of the cell types
}
\details{
The cell type contributions are visualized as a heatmap
}
\author{
Pavlo Lutsik
}

