% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.plot.marker.fstat}
\alias{rnb.plot.marker.fstat}
\title{rnb.plot.marker.fstat}
\usage{
rnb.plot.marker.fstat(ct.object, writeToFile = FALSE, ...)
}
\arguments{
\item{ct.object}{Object of class \code{CellTypeInferenceResult} as returned by \link{rnb.execute.ct.estimation}.}

\item{writeToFile}{If \code{TRUE}, the plot will be written to a file.}

\item{...}{Other arguments to \code{\link{createReportPlot}}.}
}
\value{
if \code{writeToFile=TRUE} an object of class \code{\linkS4class{ReportPlot}},
						and the plotted reordered F-statistics vector otherwise
}
\description{
Plot the the cell type marker selection based on the reference methylome data
}
\details{
The F-statistic values from the cell type association model (first part of eqn. (1) in [1]) are plotted in decreasing order
			for all tested CpG positions. A vertical line gives a cut-off for the number of selected cell type markers.
}
\author{
Pavlo Lutsik
}
\references{
1. Houseman, Eugene and Accomando, William and Koestler, Devin and Christensen, Brock and Marsit, Carmen and Nelson,
	Heather and Wiencke, John and Kelsey, Karl. DNA methylation arrays as surrogate measures of cell mixture distribution. BMC Bioinformatics 2012, 13:86
}

