% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlPlotsBiSeq.R
\name{rnb.plot.num.sites.covg}
\alias{rnb.plot.num.sites.covg}
\title{rnb.plot.num.sites.covg}
\usage{
rnb.plot.num.sites.covg(rnbs, addSampleNames = (length(samples(rnbs)) < 100),
  bar.percentiles = c(0.25, 0.75))
}
\arguments{
\item{rnbs}{RnBiseqSet object}

\item{addSampleNames}{should the sample names be added to the plot}

\item{bar.percentiles}{the percentiles to be used for the error bars. Must be a vector of length 2 of which the
first two elements will be used}
}
\value{
plot as an object of type \code{\link{ggplot}}
}
\description{
plot the number of sites vs median and other percentiles of coverage
}
\author{
Fabian Mueller
}

