% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.remove.annotation}
\alias{rnb.remove.annotation}
\title{rnb.remove.annotation}
\usage{
rnb.remove.annotation(type, assembly = "hg19")
}
\arguments{
\item{type}{One-element \code{character} vector giving the name of the region annotation.}

\item{assembly}{Genome assembly of interest. See \code{\link{rnb.get.assemblies}} for the list of supported genomes.}
}
\value{
Invisibly, \code{TRUE} if the annotation has been successfully deleted, or \code{FALSE} if the specified
        region type is not supported.
}
\description{
Deletes a region annotation table. Use this function with caution; its operation cannot be undone.
}
\examples{
\dontrun{
t.regions <- rnb.get.annotation("tiling")
rnb.remove.annotation("tiling")
}
}
\author{
Fabian Mueller
}
\seealso{
\code{\link{rnb.get.annotation}}, \code{\link{rnb.region.types}}
}

