% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rnb.run.dj}
\alias{rnb.run.dj}
\title{rnb.run.dj}
\usage{
rnb.run.dj()
}
\value{
Nothing of particular interest
}
\description{
Starts the RnBeads Data Juggler (RnBeadsDJ) for configuring and running RnBeads analyses from the web browser
}
\details{
A Shiny app is launched in the web browser
}
\seealso{
\code{\link{rnb.run.analysis}} for starting an analysis pipeline
}
\author{
Fabian Mueller
}
