% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.run.example}
\alias{rnb.run.example}
\title{rnb.run.example}
\usage{
rnb.run.example(index = 4L, dir.output = "example")
}
\arguments{
\item{index}{Example to start. This must be one of \code{1}, \code{2}, \code{3} or \code{4}.}

\item{dir.output}{One-element \code{character} vector specifying the directory to contain the downloaded data files
and generated reports. This must be a non-existent path, as this function attempts to create it.}
}
\value{
Invisibly, the loaded, normalized and/or possibly filtered dataset as an object of type inheriting
        \code{\linkS4class{RnBSet}}.
}
\description{
Executes the analysis pipeline for an example from the RnBeads web site.
}
\details{
For more information about the examples, please visit the dedicated
\href{http://rnbeads.mpi-inf.mpg.de/examples.php}{page on the RnBeads web site}.
}
\examples{
\dontrun{
rnb.run.example()
}
}
\author{
Yassen Assenov
}
\seealso{
\code{\link{rnb.run.analysis}} for starting the analysis pipeline from a local data source
}

