% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xenomix.R
\name{rnb.run_xeno}
\alias{rnb.run_xeno}
\title{run_xeno}
\usage{
rnb.run_xeno(idat_path, n_cores = 1)
}
\arguments{
\item{idat_path}{Path to the idat file for which the mouse fraction is to be computed}

\item{n_cores}{The number of cores to be used}
}
\value{
a \code{data.frame} with the estimated mouse content
}
\description{
Run Xeno Mouse Fraction Analysis
}
\author{
Nima Esmaeelpour
}
