% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.sample.replicates}
\alias{rnb.sample.replicates}
\title{rnb.sample.replicates}
\usage{
rnb.sample.replicates(rnb.set, replicate.id.col)
}
\arguments{
\item{rnb.set}{Methylation dataset as an object of type inheriting \code{\linkS4class{RnBSet}}.}

\item{replicate.id.col}{Trait (column name in the sample annotation table) that indicates sample replicates.
   Replicates should have the same value for this trait, while samples without replicates are
expected to have unique values or missing values.}
}
\value{
List of length of the number of replicates in the dataset. Each element is an \code{integer} vector storing
        the corresponding sample indices.
}
\description{
Identifies sample replicates defined in the given sample annotation table.
}
\author{
Fabian Mueller
}

