% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialVariability.R
\name{rnb.section.diffVar.region}
\alias{rnb.section.diffVar.region}
\title{rnb.section.diffVar.region}
\usage{
rnb.section.diffVar.region(rnb.set, diff.meth, report, gzTable = FALSE,
  level = 1)
}
\arguments{
\item{rnb.set}{Object of type \code{\linkS4class{RnBSet}} containing methylation information}

\item{diff.meth}{RnBDiffMeth object. See \code{\link{RnBDiffMeth-class}} for details.}

\item{report}{Report object to which the content is added}

\item{gzTable}{Flag indicating if tables should be gzipped}

\item{level}{Which level of section should be created. See \code{rnb.add.section}.}
}
\value{
The modified report object
}
\description{
Adds information for differentially variable regions to the report.
}
\author{
Michael Scherer
}
