% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.set.annotation.and.cpg.stats}
\alias{rnb.set.annotation.and.cpg.stats}
\title{rnb.set.annotation.and.cpg.stats}
\usage{
rnb.set.annotation.and.cpg.stats(type, regions, description = NULL,
  assembly = "hg19")
}
\arguments{
\item{type,description,assembly}{Parameters handled exactly as in \code{\link{rnb.set.annotation}}}

\item{regions}{a data.frame handled similarly as by \code{\link{rnb.set.annotation}} with the exception that
the genomic location columns should be specified using upper case first letters}
}
\description{
wrapper for \code{\link{rnb.set.annotation}} to accept the region format as output by \code{annotation(rnb.set)}.
Additionally, CpG statistics are added to the annotation.
}
\author{
Fabian Mueller
}
\seealso{
\code{\link{rnb.set.annotation}}
}

