% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.show.report}
\alias{rnb.show.report}
\title{rnb.show.report}
\usage{
rnb.show.report(report)
}
\arguments{
\item{report}{\code{\linkS4class{Report}} object to open.}
}
\description{
Opens the given HMTL report file in the browser.
}
\author{
Pavlo Lutsik
}

