% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{rnb.step.ageprediction}
\alias{rnb.step.ageprediction}
\title{rnb.step.ageprediction}
\usage{
rnb.step.ageprediction(object, report)
}
\arguments{
\item{object}{a \code{\linkS4class{RnBeadSet}} object}

\item{report}{Report in which the age prediction section should be created. This must be an object of
type \code{\linkS4class{Report}}.}
}
\value{
the modified report object
}
\description{
Creates section for the age prediction part in the report
}
\author{
Michael Scherer
}

