% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{rnb.step.cnv}
\alias{rnb.step.cnv}
\title{rnb.step.cnv}
\usage{
rnb.step.cnv(rnb.set, report)
}
\arguments{
\item{rnb.set}{An object of type \code{\linkS4class{RnBeadRawSet}}}

\item{report}{Report on quality control to contain the generated sections. This must be an object of type
\code{\linkS4class{Report}}.}
}
\value{
The modified report.
}
\description{
Performs copy number calling from the Infinium intenstity data and adds the results to the report
}
\author{
Pavlo Lutsik
}
