% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{sampleMethApply,RnBSet-method}
\alias{sampleMethApply}
\alias{sampleMethApply,RnBSet-method}
\title{sampleMethApply-methods}
\usage{
\S4method{sampleMethApply}{RnBSet}(object, fn, type = "sites", ...)
}
\arguments{
\item{object}{object inheriting from \code{\linkS4class{RnBSet}}}

\item{fn}{function to be applied}

\item{type}{\code{character} singleton. Specify "sites" (default) or a region type over which the function is applied}

\item{...}{arguments passed on to the function}
}
\value{
Result analogous to \code{apply(meth(rnbSet, type), 2, FUN=FUN)}
}
\description{
Applies a function over the methylation values for all samples in an \code{RnBSet} using a low memory footprint.
}
\seealso{
\code{\link[=meth,RnBSet-method]{meth}} Retrieving the matrix of methylation values
}

