% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R, R/clustering.R
\docType{methods}
\name{samples,RnBSet-method}
\alias{samples}
\alias{samples,RnBSet-method}
\alias{samples,RnBeadClustering-method}
\title{samples-methods}
\usage{
\S4method{samples}{RnBSet}(object)

\S4method{samples}{RnBeadClustering}(object)
}
\arguments{
\item{object}{Dataset of interest.}
}
\value{
\code{character} vector of sample identifiers.
}
\description{
Extracts sample identifiers
}
\details{
The column of the sample annotation table which contain identifiers is globally controlled via the
 \code{"identifiers.column"} option. In case the latter is \code{NULL} column names of the matrix returned
by the \code{meth} method are treated as sample identifiers. In case the latter are also missing, a \code{character}
vector with sample numbers is returned.
}
\examples{
\donttest{
library(RnBeads.hg19)
data(small.example.object)
samples(rnb.set.example)
}
}

