% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\name{save.rnb.diffmeth}
\alias{save.rnb.diffmeth}
\title{save.rnb.diffmeth}
\usage{
save.rnb.diffmeth(object, path)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}

\item{path}{path on the disk to save to.}
}
\description{
save an \code{\linkS4class{RnBDiffMeth}} object to disk
}
\author{
Fabian Mueller
}

