% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\name{save.rnb.set}
\alias{save.rnb.set}
\title{save.rnb.set}
\usage{
save.rnb.set(object, path, archive = TRUE)
}
\arguments{
\item{object}{\code{RnBSet}-inheriting object.}

\item{path}{the name of the output file (or directory if \code{archive} is \code{FALSE})
without an extension. If only the file name is given the object will be saved
in the current working directory.}

\item{archive}{if \code{TRUE} (default value) the output is a ZIP-file.}
}
\value{
invisibly, the full path to the ZIP file (if \code{archive} is \code{TRUE}),
					  or to the output directory (otherwise)
}
\description{
Consistent saving of an \code{RnBSet} objects with large matrices of type \link{ff}.
}
\details{
The saved object can be reloaded with the \link{load.rnb.set} function.
}
\author{
Pavlo Lutsik
}

