% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBDiffMeth-class.R
\docType{methods}
\name{save.tables,RnBDiffMeth-method}
\alias{save.tables,RnBDiffMeth-method}
\alias{save.tables}
\title{save.tables-methods}
\usage{
\S4method{save.tables}{RnBDiffMeth}(object, file)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}

\item{file}{path on the disk to save to.}
}
\value{
success
}
\description{
save the disk dumped tables to an ff archive for later reloading
}
\examples{
\donttest{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
pcols <- c("Sample_Group","Treatment")
tdir <- tempfile()
dm <- rnb.execute.computeDiffMeth(rnb.set.example,pcols,disk.dump=TRUE,disk.dump.dir=tdir)
save.tables(dm,tempfile())
}
}
\author{
Fabian Mueller
}
