% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{setExecutable,ClusterArchitecture,character,character-method}
\alias{setExecutable}
\alias{setExecutable,ClusterArchitecture,character,character-method}
\title{setExecutable-methods}
\usage{
\S4method{setExecutable}{ClusterArchitecture,character,character}(object,
  exec.name, exec.loc)
}
\arguments{
\item{object}{\code{\linkS4class{ClusterArchitecture}} object}

\item{exec.name}{A name/identifier that will be associated with the given executable}

\item{exec.loc}{The executable's location}
}
\value{
The modified object
}
\description{
Tells the cluster architecture about an executable that can be submitted as job
}
\author{
Fabian Mueller
}

