% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{setModuleNumCores,RnBClusterRun,integer,character-method}
\alias{setModuleNumCores}
\alias{setModuleNumCores,RnBClusterRun,integer,character-method}
\title{setModuleNumCores-methods}
\usage{
\S4method{setModuleNumCores}{RnBClusterRun,integer,character}(object, num.cores,
  modules = "all")
}
\arguments{
\item{object}{\code{\linkS4class{RnBClusterRun}} object}

\item{num.cores}{an integer specifying the number of cores to be used}

\item{modules}{vector of applicable pipeline modules. Can be \code{"all"} to specify all modules}
}
\value{
The modified object
}
\description{
Specifies the number of cores used by the different pipeline modules
}
\author{
Fabian Mueller
}

