% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{setModuleResourceRequirements,RnBClusterRun,character,character-method}
\alias{setModuleResourceRequirements}
\alias{setModuleResourceRequirements,RnBClusterRun,character,character-method}
\title{setModuleResourceRequirements-methods}
\usage{

  \S4method{setModuleResourceRequirements}{RnBClusterRun,character,character}(object,
  resources, modules = "all")
}
\arguments{
\item{object}{\code{\linkS4class{RnBClusterRun}} object}

\item{resources}{A NAMED character vector containing the resource reuirements as value and the resource name as name}

\item{modules}{vector of applicable pipeline modules. Can be \code{"all"} to specify all modules}
}
\value{
The modified object
}
\description{
Specifies resource requirements for the different pipeline modules
}
\author{
Fabian Mueller
}

