% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{simpleGlmnet}
\alias{simpleGlmnet}
\title{simpleGlmnet}
\usage{
simpleGlmnet(trainRnBSet, filePath = "")
}
\arguments{
\item{trainRnBSet}{An \code{RnBeadSet} object containing the methylation info on which 
the new predictor should be trained}

\item{path}{Path in which the new predictor should be written}
}
\value{
Absolute path to the corresponding predictor. Null if the function was 
	unable to create such an predictor.
}
\description{
This function actually fits the metylation data to the available age data by fitting
a general regularized linear model (alpha parameter = 0.8) and then building a linear
regression model upon the coefficients that had a non-zero coefficient in the glm. Specifically
dedicated to BeadChip data.
}

