% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{simpleGlmnetBiseq}
\alias{simpleGlmnetBiseq}
\title{simpleGlmnetBiseq}
\usage{
simpleGlmnetBiseq(trainRnBSet, filePath = "")
}
\arguments{
\item{trainRnBSet}{a \code{RnBiseqSet} object containing the methylation info on which 
the new predictor should be trained}

\item{filePath}{path to which the new predictor should be written out}
}
\value{
absolute path to the corresponding predictor. \code{NULL} if the function was 
	unable to create such an predictor.
}
\description{
This function fits the metylation data to the available ages by fitting
a generalized linear model (alpha parameter = 0.8). The lambda parameter is chosen by
10-fold cross-validation. This function was created for sequencing-based data.
}
\author{
Michael Scherer
}

