% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{simpleGlmnetEvaluate}
\alias{simpleGlmnetEvaluate}
\title{simpleGlmnetEvaluate}
\usage{
simpleGlmnetEvaluate(methData, ages)
}
\arguments{
\item{methData}{input methylation data as a data.frame for age prediction}

\item{ages}{training ages}
}
\value{
the age prediction function to be applied in each fold
}
\description{
This function is needed to perform cross-validation. In contrast to simpleGlmnet
it does not write the predictor to a csv-file but returns a prediction 
function that is to be applied in each fold.
}
\author{
Michael Scherer
}

