% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigFf.R
\docType{methods}
\name{[<-,BigFfMat,ANY,ANY,ANY-method}
\alias{[<-,BigFfMat,ANY,ANY,ANY-method}
\title{Replace parts of BigFfMat}
\usage{
\S4method{[}{BigFfMat,ANY,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{BigFfMat object}

\item{i}{row indices (integer, logical, character are allowed)}

\item{j}{column indices (integer, logical, character are allowed)}

\item{value}{replacement values}
}
\description{
Replace parts of BigFfMat
}
