% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{trainPredictor}
\alias{trainPredictor}
\title{trainPredictor}
\usage{
trainPredictor(rnbSet, data.dir)
}
\arguments{
\item{rnbSet}{An \code{RnBSet} object containing the methylation info on which 
the new predictor should be trained}

\item{data.dir}{directory to which the resulting predictor should be written out}
}
\value{
absolute path to the corresponding predictor. The path is also set as the 
	option \code{inference.age.prediction.predictor}
}
\description{
This function is the starting point for training a new predictor based on the the input dataset.
}
\author{
Michael Scherer
}

