% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{updateMethylationSites,RnBSet-method}
\alias{updateMethylationSites,RnBSet-method}
\alias{updateMethylationSites}
\title{updateMethylationSites-methods}
\usage{
\S4method{updateMethylationSites}{RnBSet}(object, meth.data, verbose = FALSE)
}
\arguments{
\item{object}{Dataset of interest.}

\item{meth.data}{This object has to be a \code{data.frame} of equal dimension than the one already contained in 
\code{object}, containing the methylation info that should be associated with the object.}

\item{verbose}{if \code{TRUE} additional diagnostic output is generated}
}
\value{
The modified dataset.
#'
}
\description{
Replaces the methylation info with the specified data frame.
}
