% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{updateRegionSummaries,RnBSet-method}
\alias{updateRegionSummaries}
\alias{updateRegionSummaries,RnBSet-method}
\title{updateRegionSummaries}
\usage{
\S4method{updateRegionSummaries}{RnBSet}(object)
}
\arguments{
\item{object}{Dataset of interest.}
}
\value{
Sample annotation information available for the dataset in the form of a \code{data.frame}.
}
\description{
Updates the region information present in an RnBSet by invoking summarize.regions on all region types
present in the object
}

