% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{writePredictorToCsv}
\alias{writePredictorToCsv}
\title{writePredictorToCsv}
\usage{
writePredictorToCsv(linearModel, path)
}
\arguments{
\item{linearModel}{Linear Model which should be stored as a csv file}

\item{path}{Path in which the new predictor should be written}
}
\value{
TRUE if the writing was successful
}
\description{
This function writes the coefficients of a linear model into a csv file.
}
\author{
Michael Scherer
}

