% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{add.age.histogram}
\alias{add.age.histogram}
\title{add.age.histogram}
\usage{
\method{add}{age.histogram}(report, ages)
}
\arguments{
\item{report}{The report object to be modified}

\item{ages}{The ages for which the distribution plots should be created}
}
\value{
Modified report object with the age histogram
}
\description{
This function is creates an age distribution plot for the predicted ages in the case
where there is no age annotation available for comparison with predicted ages
}
\author{
Michael Scherer
}

