% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{add.agecomparison.plot}
\alias{add.agecomparison.plot}
\title{add.agecomparison.plot}
\usage{
\method{add}{agecomparison.plot}(report, object, actualAges, predictedAges)
}
\arguments{
\item{report}{report in which the corresponding plot should be integrated.}

\item{object}{methylation dataset as an object of type \code{\linkS4class{RnBSet}}.}

\item{actualAges}{annotated ages as found in the sample annotation sheet.}

\item{predictedAges}{ages as predicted by the age prediction algorithm.}
}
\value{
modified report with the age plots integrated.
}
\description{
Adds a plot that compares predicted ages by the age predictor with the annotated ages of the samples. The individual
points are marked in different colors and shapes depending on different samples characteristics as defined by
\code{rnb.sample.groups}.
}
\author{
Michael Scherer
}

