% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{add.combination.plot}
\alias{add.combination.plot}
\title{add.combination.plot}
\usage{
\method{add}{combination.plot}(report, object, actualAges, predictedAges)
}
\arguments{
\item{report}{report in which the corresponding plot should be integrated.}

\item{object}{methylation dataset as an object of type \code{\linkS4class{RnBSet}}.}

\item{actualAges}{annotated ages as found in the sample annotation sheet.}

\item{predictedAges}{ages as predicted by the age prediction algorithm.}
}
\value{
modified report with the combination plot integrated.
}
\description{
This function adds a plot that is the combination of two plots in one plot.The upper 
panel shows the distribution and the quantiles for the differences between predicted and annotated
ages and the lower panel shows the individual errors for each sample.
}
\author{
Michael Scherer
}

