% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{add.error.plot}
\alias{add.error.plot}
\title{add.error.plot}
\usage{
\method{add}{error.plot}(report, object, actualAges, predictedAges)
}
\arguments{
\item{report}{report in which the corresponding plot should be integrated.}

\item{object}{methylation dataset as an object of type \code{\linkS4class{RnBSet}}.}

\item{actualAges}{annotated ages as found in the sample annotation sheet.}

\item{predictedAges}{ages as predicted by the age prediction algorithm.}
}
\value{
modified report with the error plot integrated.
}
\description{
Adds a plot to the given report that describes the differences between predicted ages by the age prediction
algorithm and the ages as annotated in the sample annotation sheet(if available). Possible outliers are marked by
their sample names.
}
\author{
Michael Scherer
}

