% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{age.anti.transformation}
\alias{age.anti.transformation}
\title{age.anti.transformation}
\usage{
age.anti.transformation(x, adult.age = 20)
}
\arguments{
\item{x}{Input age about to be transformed back}

\item{adult.age}{Threshold for deciding between 'fast' and 'slow' changes in DNA methylation}
}
\value{
Backtransformed age
}
\description{
This function is used to transform the values back to actual ages, that are corresponding
to the predicted ages by the prediction algorithm
}
\details{
The function is directly taken from Steve Horvath's prediction algorithm
DNA methylation age of human tissues and cell types, Steve Horvath, Genome Biology, 2013
}
\author{
Steve Horvath
}

