% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{age.transformation}
\alias{age.transformation}
\title{age.transformation}
\usage{
age.transformation(x, adult.age = 20)
}
\arguments{
\item{x}{input age about to be transformed}

\item{adult.age}{threshold for deciding between 'fast' and 'slow' changes in DNA methylation}
}
\value{
transformed age
}
\description{
This function is used to transform the input ages, to account for the fact that
the human DNA methylation pattern changes more in the first phase of life than in the later.
}
\details{
The function is directly taken from Steve Horvath's prediction algorithm
DNA methylation age of human tissues and cell types, Steve Horvath, Genome Biology, 2013
}
\author{
Steve Horvath
}

