% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.RnBeadRawSet}
\alias{as.RnBeadRawSet}
\title{as("MethyLumiSet", "RnBeadRawSet")}
\description{
Convert a \code{\linkS4class{MethyLumiSet}} object to \code{\linkS4class{RnBeadRawSet}}

Convert a \code{\linkS4class{RnBeadRawSet}} object to \code{\linkS4class{MethyLumiSet}}
}

