% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{assembly,RnBSet-method}
\alias{assembly}
\alias{assembly,RnBSet-method}
\title{assembly-methods}
\usage{
\S4method{assembly}{RnBSet}(object)
}
\arguments{
\item{object}{Dataset of interest.}
}
\value{
Sample annotation information available for the dataset in the form of a \code{data.frame}.
}
\description{
Extracts information about assembly
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
assembly(rnb.set.example) # "hg19"
}
}

