% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{covg,RnBSet-method}
\alias{covg}
\alias{covg,RnBSet-method}
\title{covg-methods}
\usage{
\S4method{covg}{RnBSet}(object, type = "sites", row.names = FALSE)
}
\arguments{
\item{object}{Dataset of interest.}

\item{type}{\code{character} singleton. If \code{sites} DNA methylation information per each available
site is returned. Otherwise should be one of region types for for which the summarized
coverage information is available}

\item{row.names}{Flag indicating of row names are to be generated in the result.}
}
\value{
coverage information available for the dataset in the form of a \code{matrix}.
}
\description{
Extract coverage information from an object of \code{RnBSet} class.
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
## per-site beta-value matrix
cvg<-covg(rnb.set.example, row.names=TRUE)
head(cvg)
}
}

