% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{cv.array}
\alias{cv.array}
\title{cv.array}
\usage{
cv.array(rnbSet)
}
\arguments{
\item{rnbSet}{RnBSet object on which the cross validation should be perfomed}
}
\value{
the result of the cross validation in a data.frame format
}
\description{
This function calls the general cross validation function from the corresponding
RnBSet object in the case of array data
}
\author{
Michael Scherer
}

