% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{get.site.test.method,RnBDiffMeth-method}
\alias{get.site.test.method}
\alias{get.site.test.method,RnBDiffMeth-method}
\title{get.site.test.method-methods}
\usage{
\S4method{get.site.test.method}{RnBDiffMeth}(object)
}
\arguments{
\item{object}{RnBDiffMeth object}
}
\value{
character describing the site test method
}
\description{
Gets the site testing method used to obtain the p-values in the differential methylation tables
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
dm <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group","Treatment"))
get.site.test.method(dm)
}
}
\author{
Fabian Mueller
}

