% Generated by roxygen2 (4.0.2): do not edit by hand
\name{greedycut.get.submatrix}
\alias{greedycut.get.submatrix}
\title{greedycut.get.submatrix}
\usage{
greedycut.get.submatrix(mm, filter.info, it.num = nrow(filter.info) -
  as.integer(1))
}
\arguments{
\item{mm}{Data \code{matrix} to be filtered.}

\item{filter.info}{Information on the filtering iterations as a \code{data.frame} returned by
\code{\link{greedycut.filter.matrix}}.}

\item{it.num}{Number of iterations to execute. Defaults to all iterations.}
}
\value{
Data matrix containing subsets of the rows and columns of \code{mm}.
}
\description{
Filters a data matrix executing the given number of iterations of Greedycut.
}
\author{
Yassen Assenov
}

