% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,ClusterArchitecture-method}
\alias{initialize,ClusterArchitecture-method}
\title{initialize.ClusterArchitecture}
\usage{
\S4method{initialize}{ClusterArchitecture}(.Object,
  name = "ClusterArchitecture")
}
\arguments{
\item{name}{A name or identifier}
}
\description{
Initialize an ClusterArchitecture object
}
\author{
Fabian Mueller
}

