% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,ClusterArchitectureSGE-method}
\alias{initialize,ClusterArchitectureSGE-method}
\title{initialize.ClusterArchitectureSGE}
\usage{
\S4method{initialize}{ClusterArchitectureSGE}(.Object,
  name = "ClusterArchitectureSGE", ...)
}
\arguments{
\item{name}{A name or identifier}

\item{...}{arguments passed on to the constructor of \code{\linkS4class{ClusterArchitecture}} (the parent class)}
}
\description{
Initialize an ClusterArchitecture object for a Sun Grid Engine (SGE)
}
\author{
Fabian Mueller
}

