% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,RnBClusterRun-method}
\alias{initialize,RnBClusterRun-method}
\title{initialize.RnBClusterRun}
\usage{
\S4method{initialize}{RnBClusterRun}(.Object, architecture)
}
\arguments{
\item{architecture}{A \code{\linkS4class{ClusterArchitecture}} object managing the settings for a scientific compute cluster.}
}
\description{
Initialize an RnBClusterRun object
}
\author{
Fabian Mueller
}

