% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBeadRawSet-class.R
\name{intensities.by.color}
\alias{intensities.by.color}
\title{intensities.by.color}
\usage{
intensities.by.color(raw.set, address.rownames = TRUE, add.oob = TRUE,
  add.controls = TRUE, add.missing = TRUE)
}
\arguments{
\item{raw.set}{RnBeadRawSet object}

\item{address.rownames}{if \code{TRUE} the rows of the returned matrices are named with the with the correspoding Illumina probe addresses}

\item{add.oob}{if \code{TRUE} the "out-of-band" intensities are included}

\item{add.controls}{if \code{TRUE} the control probe intensities are included}

\item{add.missing}{if \code{TRUE} the rows for the probes missing in \code{raw.set} is imputed with \code{NA} values}
}
\value{
a \code{list} with elements \code{Cy3} and \code{Cy5} containing average bead intensities 
measured for each probe in the green and red channels, respectively
}
\description{
Rearranges information from "M" and "U" slots of a RnBeadsRawSet object by color channer.
}
\author{
Pavlo Lutsik
}

