% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{is.valid,RnBDiffMeth-method}
\alias{is.valid}
\alias{is.valid,RnBDiffMeth-method}
\title{is.valid-methods}
\usage{
\S4method{is.valid}{RnBDiffMeth}(object, verbose = FALSE)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}

\item{verbose}{print more info to the logger}
}
\value{
TRUE iff all differential methylation tables are present and accounted for
}
\description{
Validate an RnBDiffMeth object, ie. verify that all differential methylation tables are specified
and accounted for
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
dm1 <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group"),region.types=c("genes","tiling"))
dm2 <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group","Treatment"),region.types=c("promoters"))
dm.join1 <- join.diffMeth(dm1,dm2)
#the following joint object is invalid, because some region type - comparison combinations are missing
is.valid(dm.join1)
dm3 <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Treatment"),region.types=c("genes","tiling"))
dm.join2 <- join.diffMeth(dm.join1,dm3)
#after joining the missing information, the new object is valid
is.valid(dm.join2)
}
}
\author{
Fabian Mueller
}

