% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCluster.R
\name{logger.machine.name}
\alias{logger.machine.name}
\title{logger.machine.name}
\usage{
logger.machine.name()
}
\value{
None (invisible \code{NULL}).
}
\description{
Log the machine name the analysis is run on
}
\author{
Fabian Mueller
}

