% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelProcessing.R
\name{parallel.getNumWorkers}
\alias{parallel.getNumWorkers}
\title{parallel.getNumWorkers}
\usage{
\method{parallel}{getNumWorkers}()
}
\value{
Number of workers used for parallel processing; \code{-1} if parallel processing is not enabled.
}
\description{
Gets the number of workers used for parallel processing.
}
\examples{
\donttest{
parallel.getNumWorkers()
parallel.setup(2)
parallel.getNumWorkers()
parallel.teardown()
parallel.getNumWorkers()
}
}
\author{
Fabian Mueller
}

