% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parallel.getNumWorkers}
\alias{parallel.getNumWorkers}
\title{parallel.getNumWorkers}
\usage{
parallel.getNumWorkers()
}
\description{
Return the number of workers used for parallel processing.
}
\examples{
\dontrun{
parallel.getNumWorkers()
parallel.setup(2)
parallel.getNumWorkers()
parallel.disable()
parallel.getNumWorkers()
}
}
\author{
Fabian Mueller
}

