% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parallel.isEnabled}
\alias{parallel.isEnabled}
\title{parallel.isEnabled}
\usage{
parallel.isEnabled()
}
\value{
\code{TRUE} if multicore processing is enabled, \code{FALSE} otherwise.
}
\description{
Checks if whether parallel processing is enabled.
}
\examples{
\dontrun{
parallel.isEnabled()
parallel.setup(2)
parallel.isEnabled()
parallel.disable()
parallel.isEnabled()
}
}
\author{
Fabian Mueller
}

