% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parallel.setup}
\alias{parallel.setup}
\title{parallel.setup}
\usage{
parallel.setup(...)
}
\arguments{
\item{...}{Parameters for \code{registerDoParallel} from the \pkg{doParallel} package.
This allows, for instance, for specificying the number of workers.}
}
\value{
\code{TRUE} (invisible) to indicate that parallelization is set up.
}
\description{
Sets up parallel processing. Requires the \pkg{foreach} and \pkg{doParallel} packages
}
\note{
Requires the packages \pkg{foreach} and \pkg{doParallel}.
}
\examples{
\dontrun{
parallel.setup(2)
}
}
\author{
Fabian Mueller
}

