% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelProcessing.R
\name{parallel.teardown}
\alias{parallel.teardown}
\title{parallel.teardown}
\usage{
\method{parallel}{teardown}()
}
\value{
\code{TRUE}, invisibly.
}
\description{
Disables parallel processing.
}
\examples{
\donttest{
parallel.getNumWorkers()
parallel.setup(2)
parallel.getNumWorkers()
parallel.teardown()
parallel.getNumWorkers()
}
}
\author{
Fabian Mueller
}

