% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{pheno,RnBSet-method}
\alias{pheno}
\alias{pheno,RnBSet-method}
\title{pheno-methods}
\usage{
\S4method{pheno}{RnBSet}(object)
}
\arguments{
\item{object}{Dataset of interest.}
}
\value{
Sample annotation information available for the dataset in the form of a \code{data.frame}.
}
\description{
Extracts sample phenotype and/or processing information.
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
pheno(rnb.set.example)
}
}

