% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{rnb.execute.age.prediction}
\alias{rnb.execute.age.prediction}
\title{rnb.execute.age.prediction}
\usage{
rnb.execute.age.prediction(object)
}
\arguments{
\item{object}{a \code{\linkS4class{RnBSet}} object for which age prediction should
be performed}
}
\value{
modified  \code{\linkS4class{RnBSet}} object
}
\description{
Performs age prediction by either the specified predictor in the option \code{inference.age.prediction.predictor}
or by the corresponding predefined predictor.
}
\author{
Michael Scherer
}

