% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\name{rnb.execute.gender.prediction}
\alias{rnb.execute.gender.prediction}
\title{rnb.execute.gender.prediction}
\usage{
rnb.execute.gender.prediction(rnb.set)
}
\description{
Deprecated function name, now called \code{\link{rnb.execute.sex.prediction}}.
}
\seealso{
rnb.execute.sex.prediction
}
