% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.execute.gender.prediction}
\alias{rnb.execute.gender.prediction}
\title{rnb.execute.gender.prediction}
\usage{
rnb.execute.gender.prediction(rnb.set)
}
\arguments{
\item{rnb.set}{Methylation dataset as an object of type \code{\linkS4class{RnBeadRawSet}}.}
}
\value{
The possibly modified dataset. If gender could be predicted, the sample annotation table is enriched with
two more columns - \code{"Predicted Male Probability"} and \code{"Predicted Gender"}.
}
\description{
Infers the gender of every sample in the given Infinium 450k dataset, based on average signal intensity values on
the autosomes and the sex chromosomes.
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
rnb.set.example <- rnb.execute.gender.prediction(rnb.set.example)
table(rnb.set.example[, "Predicted Gender"])
}
}
\author{
Yassen Assenov
}

