% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{rnb.execute.training}
\alias{rnb.execute.training}
\title{rnb.execute.age.training}
\usage{
rnb.execute.training(object, path = "")
}
\arguments{
\item{object}{a \code{\linkS4class{RnBeadSet}} object for which the predictor should
be trained}

\item{path}{path to which the predictor should be written}
}
\value{
the modified  \code{\linkS4class{RnBeadSet}} object
}
\description{
Trains a new predictor on the specified data set and writes it to the given path
}
\author{
Michael Scherer
}

