% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.export.annotation}
\alias{rnb.export.annotation}
\title{rnb.export.annotation}
\usage{
rnb.export.annotation(fname, type, assembly = "hg19", format = "bed")
}
\arguments{
\item{fname}{One-element \code{character} vector giving the name of the file to contain the annotation data. If
this file already exists, it will be overwritten.}

\item{type}{One-element \code{character} vector giving the name of the region annotation.}

\item{assembly}{Genome assembly of interest. See \code{\link{rnb.get.assemblies}} for the list of supported genomes.}

\item{format}{Output format. currently only \code{"bed"} is supported.}
}
\description{
Export the annotation to a defined format (currently only bed is supported
}
\examples{
\dontrun{
rnb.export.annotation(tempfile(pattern="promoters",fileext=".bed"),"promoters")
}
}
\author{
Fabian Mueller
}

