% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.get.assemblies}
\alias{rnb.get.assemblies}
\title{rnb.get.assemblies}
\usage{
rnb.get.assemblies()
}
\value{
All supported genome assemblies in the form of a \code{character} vector. These are \code{"hg19"},
        \code{"mm10"}, \code{"mm9"} and \code{"rn5"}.
}
\description{
Gets the supported genome assemblies.
}
\examples{
\dontrun{
"hg19" \%in\% rnb.get.assemblies()
}
}
\author{
Yassen Assenov
}

