% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.infinium.control.targets}
\alias{rnb.infinium.control.targets}
\title{rnb.infinium.control.targets}
\usage{
rnb.infinium.control.targets(target = "probes450")
}
\arguments{
\item{target}{A singleton of type \code{character}, specifying the microarray platform.
                  \code{"probes450"} and \code{"probes27"} correspond to HumanMethylation450
                 respectively HumanMethylation27 microarrays}
}
\value{
\code{character} vector of control targets.
}
\description{
Extracts all control probe types in the HumanMethylation450 assay.
}
\examples{
\dontrun{
"NEGATIVE" \%in\% rnb.infinium.control.targets()
}
}
\author{
Pavlo Lutsik
}

