% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.performance.profile}
\alias{rnb.performance.profile}
\title{rnb.performance.profile}
\usage{
rnb.performance.profile(profile)
}
\arguments{
\item{profile}{A character singleton specifying the option profile. Can attain values
                  "450k_minimal", "450k_moderate", "450k_full", "bs_minimal", "bs_moderate",
                  "bs_full".}
}
\description{
Enables one of the pre-installed option profiles
}
\author{
Pavlo Lutsik
}

