% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.plot.snp.barplot}
\alias{rnb.plot.snp.barplot}
\title{rnb.plot.snp.barplot}
\usage{
rnb.plot.snp.barplot(rnb.set, sample, writeToFile = FALSE,
  numeric.names = FALSE, ...)
}
\arguments{
\item{rnb.set}{\code{\linkS4class{RnBeadRawSet}} or \code{\linkS4class{RnBeadSet}} object}

\item{sample}{unique sample identifier. In case \code{rnb.getOption("identifiers.column")} is not \code{NULL},
\code{sample} should attain values from the corresponding column, or \code{colnames(meth(rnb.set))}
otherwise.}

\item{writeToFile}{flag specifying whether the output should be saved as \code{\linkS4class{ReportPlot}}}

\item{numeric.names}{if \code{TRUE} and \code{writeToFile} is \code{TRUE}substitute the plot options in the plot file name with digits}

\item{...}{other arguments to \code{\link{createReportPlot}}}
}
\value{
plot as an object of type \code{\linkS4class{ReportPlot}} if \code{writeToFile} is \code{TRUE} and of class
			\code{\link{ggplot}} otherwise.
}
\description{
Bar plots of beta-values from the genotyping probes
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
samp<-samples(rnb.set.example)[1]
rnb.plot.snp.barplot(rnb.set.example, samp)
}
}
\author{
Pavlo Lutsik
}

