% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.plot.snp.boxplot}
\alias{rnb.plot.snp.boxplot}
\title{rnb.plot.snp.boxplot}
\usage{
rnb.plot.snp.boxplot(rnb.set, writeToFile = FALSE, ...)
}
\arguments{
\item{rnb.set}{\code{\linkS4class{RnBeadSet}} object}

\item{writeToFile}{a flag specifying whether the output should be saved as \code{\linkS4class{ReportPlot}}}

\item{...}{other arguments to \code{\link{createReportPlot}}}
}
\description{
Box plots of beta-values from the genotyping probes
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
rnb.plot.snp.boxplot(rnb.set.example)
}
}
\author{
Pavlo Lutsik
}

