% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rnb.plot.snp.heatmap}
\alias{rnb.plot.snp.heatmap}
\title{rnb.plot.snp.heatmap}
\usage{
rnb.plot.snp.heatmap(rnb.set, writeToFile = FALSE, ...)
}
\arguments{
\item{rnb.set}{\code{\linkS4class{RnBeadRawSet}} or \code{\linkS4class{RnBeadSet}} object}

\item{writeToFile}{flag specifying whether the output should be saved as \code{\linkS4class{ReportPlot}}}

\item{...}{other arguments to \code{\link{createReportPlot}}}
}
\value{
plot as an object of type \code{\linkS4class{ReportPlot}} if \code{writeToFile} is \code{TRUE} and of class
			\code{\link{ggplot}} otherwise.
}
\description{
Heatmap of beta-values from genotyping probes
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
rnb.plot.snp.heatmap(rnb.set.example)
}
}
\author{
Pavlo Lutsik
}

