% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{rnb.step.ageprediction}
\alias{rnb.step.ageprediction}
\title{rnb.step.ageprediction}
\usage{
rnb.step.ageprediction(object, report)
}
\arguments{
\item{object}{a \code{\linkS4class{RnBSet}} object}

\item{report}{report in which the age prediction section should be created. This must be an object of
type \code{\linkS4class{Report}}.}
}
\value{
modified report object
}
\description{
This function is the starting point for the report generation of the epigenetic age prediction. It call the functions
\code{rnb.section.ageprediction} and, if activated, \code{rnb.run.cross.validation}.
}
\author{
Michael Scherer
}

