% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{RnBeads.data}
\alias{RnBeads.data}
\alias{hg19}
\alias{mm10}
\alias{mm9}
\alias{rn5}
\title{RnBeads Annotation Tables}
\format{\code{list} of four elements - \code{"regions"}, \code{"sites"}, \code{"controls"} and \code{"mappings"}.
        These elements are described below.
        \describe{
          \item{\code{"regions"}}{\code{list} of \code{NULL}s; the names of the elements correspond to the built-in
               region annotation tables. Once the default annotations are loaded, the attribute \code{"builtin"} is
               a \code{logical} vector storing, for each region annotation, whether it is the default (built-in) or
               custom.}
          \item{\code{"sites"}}{\code{list} of \code{NULL}s; the names of the elements correspond to the site and
               probe annotation tables.}
          \item{\code{"controls"}}{\code{list} of \code{NULL}s; the names of the elements correspond to the control
               probe annotation tables. The attribute \code{"sites"} is a \code{character} vector pointing to the
               site annotation that encompasses the respective control probes.}
          \item{\code{"mappings"}}{\code{list} of \code{NULL}s; the names of the elements correspond to the built-in
               region annotation tables.}
        }}
\description{
RnBeads uses sets of annotation tables and mappings (from regions to sites) for each of the supported genomes. The
structures for one assembly are stored in a separate dedicated data package. Currently, the following assemblies are
supported:
\describe{
  \item{\code{"hg19"}}{through the package \pkg{RnBeads.hg19}}
  \item{\code{"mm10"}}{through the package \pkg{RnBeads.mm10}}
  \item{\code{"mm9"}}{through the package \pkg{RnBeads.mm9}}
  \item{\code{"rn5"}}{through the package \pkg{RnBeads.rn5}}
}
}
\details{
The assembly-specific structures are automatically loaded upon initialization of the annotation, that is, by the
first valid call to any of the following functions: \code{\link{rnb.get.chromosomes}},
\code{\link{rnb.get.annotation}}, \code{\link{rnb.set.annotation}}, \code{\link{rnb.get.mapping}},
\code{\link{rnb.annotation.size}}. Adding an annotation amounts to attaching its table(s) and mapping structures to
the scaffold.
}
\author{
Yassen Assenov
}
\keyword{datasets}

