% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{RnBeadSet-class}
\alias{RnBeadSet-class}
\alias{initialize,RnBeadSet-method}
\title{RnBeadSet Class}
\usage{
\S4method{initialize}{RnBeadSet}(.Object, pheno, betas, p.values = NULL,
  bead.counts = NULL, platform = "450k",
  region.types = rnb.region.types.for.analysis("hg19"),
  useff = rnb.getOption("disk.dump.big.matrices"))
}
\arguments{
\item{pheno}{phenotypic data.}

\item{betas}{matrix of beta values. Should contain Infinium probe identifiers as row names.}

\item{p.values}{matrix of detection p-values.}

\item{bead.counts}{matrix of bead counts per probe.}

\item{platform}{\code{character} singleton specifying the microarray platform: \code{"450k"} corresponds to HumanMethylation450 microarray, and \code{"27k"} stands for HumanMethylation27.}

\item{region.types}{a \code{character} vector specifying the region types, for which the methylation infromation will be summarized.}

\item{useff}{if \code{TRUE} the data matrices will be stored as \code{ff} objects}
}
\description{
Stores the preprocessed information from HumanMethylation experiments

Initializes \code{RnBeadSet} object with data from supplied data matrices.
}
\details{
There are multiple ways to create an object of type \code{RnBeadSet}:
\describe{
 \item{Loading from files}{Dataset can be loaded from text or binary files. See the function
      \code{\link{rnb.execute.import}} for more details.}
 \item{Downloading from GEO}{See the function \code{\link{read.geo}} for details.}
 \item{Converting from \code{MethyLumiSet}}{...}
}

Row names of the \code{betas} matrix are used to match the data to annotation. All matrices
should have identical dimensions.
}
\section{Slots}{

\describe{
  \item{\code{pval.sites}}{\code{matrix} of detection p-values with the same dimensions as \code{betas}, or
       \code{NULL} if the detection p-values are not available.}
  \item{\code{pval.regions}}{\code{list} of methylation \code{matrix} objects, one per available region type. Every row in a
		matrix corresponds to a methylation site, and every column - to a sample.}
  \item{\code{covg.sites}}{\code{matrix} of bead counts per probe with the same dimensions as \code{betas}, or
       \code{NULL} if this data are not available.}
  \item{\code{qc}}{Quality control probe information in the form of a \code{list} of two elements - \code{"Cy3"} and
       \code{"Cy5"}, storing intensities of probes on the green and red channels, respectively. This slot's value is
       \code{NULL} if no control probe information is available.}
}
}

\section{Methods and Functions}{

\describe{
  \item{\code{\link[=samples,RnBSet-method]{samples}}}{Gets the identifiers of all samples in the dataset.}
  \item{\code{\link[=pheno,RnBSet-method]{pheno}}}{Gets the phenotypic and processing data of the dataset.}
  \item{\code{\link[=meth,RnBSet-method]{meth}}}{Gets the \code{matrix} of methylation beta-values of the dataset.}
  \item{\code{\link[=dpval,RnBeadSet-method]{dpval}}}{Gets the \code{matrix} of detection p-values of the dataset.}
  \item{\code{\link[=covg,RnBSet-method]{covg}}}{Gets the \code{matrix} of bead counts of the dataset.}
  \item{\code{\link[=qc,RnBeadSet-method]{qc}}}{Gets the intensities of the quality control probes.}
  \item{\code{\link[=remove.sites,RnBSet-method]{remove.sites}}}{Removes probes from the dataset.}
  \item{\code{\link[=remove.samples,RnBeadSet-method]{remove.samples}}}{Removes samples from the dataset.}
  \item{\code{\link[BiocGenerics]{combine}}}{Combines two datasets.}
}
}
\author{
Pavlo Lutsik
}

