% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{RnBiseqSet-class}
\alias{RnBiseqSet-class}
\alias{initialize,RnBiseqSet-method}
\title{RnBiseqSet Class}
\usage{
\S4method{initialize}{RnBiseqSet}(.Object, pheno, sites, meth.sites, covg.sites,
  assembly = "hg19", target = "CpG", regions = list(),
  meth.regions = list(), covg.regions = NULL,
  region.types = rnb.region.types.for.analysis(assembly),
  useff = rnb.getOption("disk.dump.big.matrices"), verbose = FALSE)
}
\arguments{
\item{pheno}{phenotypic data.}

\item{sites}{matrix of CpG site mappings.}

\item{meth.sites}{matrix of summarized methylation calls}

\item{covg.sites}{matrix with read coverage information}

\item{assembly}{the genome assembly}

\item{target}{target DNA methylation features (CpG sites)}

\item{regions}{list of region mapping matrices}

\item{meth.regions}{list of matrices with summarized region-level methylation calls}

\item{covg.regions}{list of matrices with summarized region-level coverage data}

\item{region.types}{region annotations for which the methylation data should be summarized}

\item{useff}{a flag specifying whether the ff functionality should be used}

\item{verbose}{a flag specifying whether the output}
}
\description{
A class for storing the DNA methylation and quality information from bisulfite sequencing experiments

Direct slot filling.
}
\details{
TBA
}
\section{Slots}{

\describe{
  \item{\code{status}}{Normalization status.}
}
}

\section{Methods and Functions}{

\describe{
  \item{\code{\link[BiocGenerics]{combine}}}{Combines two datasets.}
}
}
\author{
Pavlo Lutsik
}

