% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{run.cross.validation}
\alias{run.cross.validation}
\title{run.cross.validation}
\usage{
run.cross.validation(rnbSet, report)
}
\arguments{
\item{rnbSet}{An \code{RnBSet} object containing the methylation info on which 
the new predictor should be trained}

\item{report}{Report to which the table should be added}
}
\value{
Modified report object
}
\description{
This function performs a 10-fold cross validation to estimate the performance of a
newly trained predictor. If \code{parallel.isEnabled()}, the function perfoms the cross
validation in parallel. The function adds a table to the specified report containing 
the result of the 10-fold cross validation.
}
\author{
Michael Scherer
}

