% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnBSet-class.R
\docType{methods}
\name{sampleMethApply,RnBSet-method}
\alias{sampleMethApply}
\alias{sampleMethApply,RnBSet-method}
\title{sampleMethApply-methods}
\usage{
\S4method{sampleMethApply}{RnBSet}(object, fn, type = "sites", ...)
}
\arguments{
\item{fn}{function to be applied}
}
\value{
Result analogous to \code{apply(meth(rnbSet, type), 2, FUN=FUN)}
}
\description{
Applies a function over the methylation values for all samples in an \code{RnBSet} using a low memory footprint.
}
\seealso{
\code{\link[=meth,RnBSet-method]{meth}} Retrieving the matrix of methylation values
}

