% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{save.tables,RnBDiffMeth-method}
\alias{save.tables}
\alias{save.tables,RnBDiffMeth-method}
\title{save.tables-methods}
\usage{
\S4method{save.tables}{RnBDiffMeth}(object, file)
}
\arguments{
\item{object}{\code{\linkS4class{RnBDiffMeth}} object}

\item{file}{path on the disk to save to.}
}
\value{
success
}
\description{
save the disk dumped tables to an ff archive for later reloading
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
logger.start(fname=NA)
dm <- rnb.execute.computeDiffMeth(rnb.set.example,pheno.cols=c("Sample_Group","Treatment"),disk.dump=TRUE,disk.dump.dir=tempfile())
save.tables(dm,tempfile())
}
}
\author{
Fabian Mueller
}

