% Generated by roxygen2 (4.0.2): do not edit by hand
\name{set.covariates.ct}
\alias{set.covariates.ct}
\title{set.covariates.ct}
\usage{
set.covariates.ct(rnb.set, ct.obj)
}
\arguments{
\item{rnb.set}{The \code{RnBSet} object to which the results should be added}

\item{ct.obj}{An object of class \code{CellTypeInferenceResult} returned by \code{rnb.execute.ct.estimation}.}
}
\value{
The modified \code{RnBSet}.
}
\description{
Adds the results of cell type estimation to an RnBSet
}

