% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sites,RnBSet-method}
\alias{sites}
\alias{sites,RnBSet-method}
\title{sites-methods}
\usage{
\S4method{sites}{RnBSet}(object)
}
\arguments{
\item{object}{Dataset of interest.}
}
\value{
A matrix of type \code{integer} describing the sites, information for which is
present in the \code{object}
}
\description{
Methylation sites object information for which is present in the \code{RnBSet} object.
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
sites(rnb.set.example)
}
}

