% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePrediction.R
\name{trainPredictor}
\alias{trainPredictor}
\title{trainPredictor}
\usage{
trainPredictor(rnbSet, data.dir)
}
\arguments{
\item{rnbSet}{An \code{RnBeadSet} object containing the methylation info on which 
the new predictor should be trained}
}
\value{
Absolute path to the corresponding predictor. The path is also set as the 
	option \code{inference.age.prediction.predictor}
}
\description{
This function is the starting point for training a new predictor based on the dataset
}

