% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{U,RnBeadRawSet-method}
\alias{U}
\alias{U,RnBeadRawSet-method}
\title{U-methods}
\usage{
\S4method{U}{RnBeadRawSet}(object, row.names = FALSE)
}
\arguments{
\item{object}{Dataset of interest.}

\item{row.names}{Flag indicating whether the resulting matrix will be assigned row names}
}
\value{
\code{matrix} of the unmethylated probe intensities
}
\description{
Extract raw unmethylated probe intensity from an object of \code{RnBeadRawSet} class.
}
\examples{
\dontrun{
library(RnBeads.hg19)
data(small.example.object)
U.intensity<-U(rnb.set.example)
head(U.intensity)
}
}

